@@ -32,8 +32,8 @@
     "url": "git+https://github.com/syntest-framework/syntest-javascript.git"
   },
   "scripts": {
-    "build:framework": "npm --prefix ../syntest-framework run build",
-    "build": "tsc --build",
+    "build:framework": "npm --prefix ../syntest-core run build",
+    "build": "npm run build:compile",
     "build:compile": "tsc --build",
     "build:watch": "tsc --build --watch",
     "clean": "rm -rf .nyc_output dist node_modules",
@@ -49,28 +49,6 @@
     "test": "mocha",
     "test:coverage": "nyc mocha",
     "test:watch": "mocha --watch"
-  },
-  "devDependencies": {
-    "@babel/plugin-proposal-numeric-separator": "^7.18.6",
-    "@babel/plugin-proposal-object-rest-spread": "^7.19.4",
-    "@babel/preset-typescript": "^7.18.6",
-    "@types/chai": "^4.3.0",
-    "@types/figlet": "^1.5.4",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.11.7",
-    "@types/sinon": "^10.0.6",
-    "@types/store": "^2.0.2",
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
-    "complexity-report-html": "^0.0.3",
-    "eslint": "^8.2.0",
-    "eslint-config-prettier": "^8.3.0",
-    "prettier": "^2.4.1",
-    "sinon": "^12.0.1",
-    "typescript": "^4.4.4"
-  },
-  "engines": {
-    "node": ">=10.24.0"
   },
   "dependencies": {
     "@babel/cli": "^7.19.3",
@@ -89,26 +67,45 @@
     "@babel/preset-react": "^7.18.6",
     "@babel/register": "^7.18.9",
     "@istanbuljs/schema": "0.1.3",
-    "@syntest/framework": "file:../syntest-framework",
-    "@types/babel-core": "^6.25.7",
-    "babel-preset-atomic": "^5.0.0",
-    "babel-preset-gatsby": "^2.24.0",
-    "chai": "^4.3.4",
-    "chai-as-promised": "^7.1.1",
-    "chalk": "^4.0.0",
-    "clear": "^0.1.0",
-    "esm": "^3.2.25",
-    "etch": "0.9.0",
-    "figlet": "^1.5.2",
-    "fs-extra": "^10.0.0",
-    "istanbul-lib-instrument": "^5.1.0",
-    "lodash": "^4.17.21",
-    "mocha": "^9.1.3",
-    "original-require": "^1.0.1",
-    "regenerator-runtime": "^0.13.9",
-    "rxjs": "^7.5.6",
-    "store": "^2.0.12",
-    "ts-babel-node": "^1.1.1",
-    "ts-node": "10.8.1"
+    "@syntest/core": "0.2.0",
+    "chai": "4.3.7",
+    "chai-as-promised": "7.1.1",
+    "chalk": "4.1.1",
+    "clear": "0.1.0",
+    "figlet": "1.5.2",
+    "fs-extra": "11.1.0",
+    "istanbul-lib-instrument": "5.1.0",
+    "lodash.clonedeep": "4.5.0",
+    "mocha": "10.2.0",
+    "original-require": "1.0.1",
+    "regenerator-runtime": "0.13.9"
+  },
+  "devDependencies": {
+    "@commitlint/cli": "17.4.2",
+    "@commitlint/config-conventional": "17.4.2",
+    "@types/babel__core": "7.1.20",
+    "@types/chai": "4.3.4",
+    "@types/figlet": "1.5.4",
+    "@types/mocha": "10.0.1",
+    "@types/node": "18.11.17",
+    "@typescript-eslint/eslint-plugin": "5.45.1",
+    "@typescript-eslint/parser": "5.45.1",
+    "commitlint": "17.4.2",
+    "eslint": "8.29.0",
+    "eslint-config-prettier": "8.5.0",
+    "husky": "8.0.3",
+    "lint-staged": "13.1.0",
+    "prettier": "2.8.1",
+    "ts-node": "10.9.1",
+    "typescript": "4.9.4"
+  },
+  "engines": {
+    "node": ">=10.24.0"
+  },
+  "lint-staged": {
+    "**/*": [
+      "npx prettier --config .prettierrc.json --ignore-path .prettierignore --write .",
+      "npx eslint --config .eslintrc.json --ignore-path .eslintignore --fix ."
+    ]
   }
 }
