@@ -1,32 +1,41 @@
 import {
   Datapoint,
   EncodingRunner,
-  ExecutionResult, getUserInterface,
+  ExecutionResult,
+  getUserInterface,
   Properties,
-} from "@syntest/framework";
+} from "@syntest/core";
 import { JavaScriptTestCase } from "../JavaScriptTestCase";
 import { JavaScriptSubject } from "../../search/JavaScriptSubject";
 import * as path from "path";
-import { JavaScriptExecutionResult, JavaScriptExecutionStatus } from "../../search/JavaScriptExecutionResult";
-import * as _ from 'lodash'
+import {
+  JavaScriptExecutionResult,
+  JavaScriptExecutionStatus,
+} from "../../search/JavaScriptExecutionResult";
+import cloneDeep = require("lodash.clonedeep");
 import { SilentMochaReporter } from "./SilentMochaReporter";
 import ExecutionInformationIntegrator from "./ExecutionInformationIntegrator";
 
 import { Runner } from "mocha";
 import { unlinkSync, writeFileSync, rmdirSync } from "fs";
 import { JavaScriptDecoder } from "../../testbuilding/JavaScriptDecoder";
-const Mocha = require('mocha')
-const originalrequire = require("original-require");
+import Mocha = require("mocha");
+import originalrequire = require("original-require");
 
 export class JavaScriptRunner implements EncodingRunner<JavaScriptTestCase> {
   protected decoder: JavaScriptDecoder;
+<<<<<<< HEAD
   protected errorProcessor: ExecutionInformationIntegrator
   protected mocha: Mocha;
+=======
+  protected errorProcessor: ExecutionInformationIntegrator;
+>>>>>>> develop
 
   constructor(decoder: JavaScriptDecoder) {
-    this.decoder = decoder
-    this.errorProcessor = new ExecutionInformationIntegrator()
+    this.decoder = decoder;
+    this.errorProcessor = new ExecutionInformationIntegrator();
 
+<<<<<<< HEAD
     if (process.listenerCount('uncaughtException') < 1) {
       process.on("uncaughtException", reason => {
         console.log(reason);
@@ -41,14 +50,23 @@ export class JavaScriptRunner implements EncodingRunner<JavaScriptTestCase> {
         // throw reason;
       });
     }
+=======
+    process.on("uncaughtException", (reason) => {
+      throw reason;
+    });
+    process.on("unhandledRejection", (reason) => {
+      throw reason;
+    });
+>>>>>>> develop
   }
 
-  async writeTestCase(filePath: string, testCase: JavaScriptTestCase, targetName: string, addLogs = false): Promise<void> {
-    const decodedTestCase = this.decoder.decode(
-      testCase,
-      targetName,
-      addLogs
-    );
+  async writeTestCase(
+    filePath: string,
+    testCase: JavaScriptTestCase,
+    targetName: string,
+    addLogs = false
+  ): Promise<void> {
+    const decodedTestCase = this.decoder.decode(testCase, targetName, addLogs);
 
     // const transpiledTestCase = ts.transpileModule(decodedTestCase, { compilerOptions: { module: ts.ModuleKind.CommonJS }}).outputText
     // await writeFileSync(filePath, transpiledTestCase);
@@ -69,16 +87,13 @@ export class JavaScriptRunner implements EncodingRunner<JavaScriptTestCase> {
     }
   }
 
-  async createSuite() {
-
-  }
-
   async run(paths: string[]): Promise<Runner> {
-    paths = paths.map((p) => path.resolve(p))
+    paths = paths.map((p) => path.resolve(p));
 
-    const argv = {
+    const argv: Mocha.MochaOptions = <Mocha.MochaOptions>(<unknown>{
       spec: paths,
       reporter: SilentMochaReporter,
+<<<<<<< HEAD
       allowUncaught: true,
       // isWorker: true,
       // parallel: true,
@@ -104,6 +119,17 @@ export class JavaScriptRunner implements EncodingRunner<JavaScriptTestCase> {
     //     // require.resolve("@babel/plugin-transform-typescript"),
     //     // require.resolve("@babel/plugin-syntax-jsx")
       ]
+=======
+    });
+
+    const mocha = new Mocha(argv); // require('ts-node/register')
+
+    // eslint-disable-next-line
+    require("regenerator-runtime/runtime");
+    // eslint-disable-next-line
+    require("@babel/register")({
+      presets: [require.resolve("@babel/preset-env")],
+>>>>>>> develop
     });
 
     for (const _path of paths) {
@@ -111,8 +137,9 @@ export class JavaScriptRunner implements EncodingRunner<JavaScriptTestCase> {
       this.mocha.addFile(_path);
     }
 
-    let runner: Runner = null
+    let runner: Runner = null;
 
+<<<<<<< HEAD
     try {
       // Finally, run mocha.
       await new Promise((resolve, reject) => {
@@ -141,16 +168,28 @@ export class JavaScriptRunner implements EncodingRunner<JavaScriptTestCase> {
     }
     // console.log('return runner');
     return runner
+=======
+    // Finally, run mocha.
+    await new Promise((resolve) => {
+      runner = mocha.run((failures) => resolve(failures));
+    });
+
+    await mocha.dispose();
+    return runner;
+>>>>>>> develop
   }
 
   async execute(
     subject: JavaScriptSubject,
     testCase: JavaScriptTestCase
   ): Promise<ExecutionResult> {
-    const testPath = path.resolve(path.join(Properties.temp_test_directory, "tempTest.spec.js"))
+    const testPath = path.resolve(
+      path.join(Properties.temp_test_directory, "tempTest.spec.js")
+    );
 
     await this.writeTestCase(testPath, testCase, subject.name);
 
+<<<<<<< HEAD
     let runner;
     try {
       runner = await this.run([testPath])
@@ -158,97 +197,104 @@ export class JavaScriptRunner implements EncodingRunner<JavaScriptTestCase> {
       console.log(e);
     }
     const stats = runner.stats
+=======
+    const runner = await this.run([testPath]);
+
+    const stats = runner.stats;
+>>>>>>> develop
 
     const test = runner.suite.suites[0]?.tests[0];
 
     // If one of the executions failed, log it
     if (stats.failures > 0) {
-      this.errorProcessor.processError(testCase, test)
+      this.errorProcessor.processError(testCase, test);
     } else {
-      this.errorProcessor.processSuccess(testCase, test)
+      this.errorProcessor.processSuccess(testCase, test);
     }
 
     // Retrieve execution traces
-    const instrumentationData = _.cloneDeep(global.__coverage__)
-    const metaData = _.cloneDeep(global.__meta__)
+    const instrumentationData = cloneDeep(global.__coverage__);
+    const metaData = cloneDeep(global.__meta__);
 
     const traces: Datapoint[] = [];
     for (const key of Object.keys(instrumentationData || {})) {
       for (const functionKey of Object.keys(instrumentationData[key].fnMap)) {
-          const fn = instrumentationData[key].fnMap[functionKey]
-          const hits = instrumentationData[key].f[functionKey]
+        const fn = instrumentationData[key].fnMap[functionKey];
+        const hits = instrumentationData[key].f[functionKey];
 
-          traces.push({
-            id: `f-${fn.line}`,
-            type: "function",
-            path: key,
-            line: fn.line,
+        traces.push({
+          id: `f-${fn.line}`,
+          type: "function",
+          path: key,
+          line: fn.line,
 
-            hits: hits,
-          })
-        }
+          hits: hits,
+        });
+      }
 
-        for (const statementKey of Object.keys(instrumentationData[key].statementMap)) {
-          const statement = instrumentationData[key].statementMap[statementKey]
-          const hits = instrumentationData[key].s[statementKey]
+      for (const statementKey of Object.keys(
+        instrumentationData[key].statementMap
+      )) {
+        const statement = instrumentationData[key].statementMap[statementKey];
+        const hits = instrumentationData[key].s[statementKey];
 
-          traces.push({
-            id: `s-${statement.start.line}`,
-            type: "statement",
-            path: key,
-            line: statement.start.line,
+        traces.push({
+          id: `s-${statement.start.line}`,
+          type: "statement",
+          path: key,
+          line: statement.start.line,
 
-            hits: hits,
-          })
-        }
+          hits: hits,
+        });
+      }
 
-        for (const branchKey of Object.keys(instrumentationData[key].branchMap)) {
-          const branch = instrumentationData[key].branchMap[branchKey]
-          const hits = instrumentationData[key].b[branchKey]
+      for (const branchKey of Object.keys(instrumentationData[key].branchMap)) {
+        const branch = instrumentationData[key].branchMap[branchKey];
+        const hits = instrumentationData[key].b[branchKey];
 
-          // if (!hits.find((h) => h !== 0)) {
-          //   // if there are no hits the meta object is not created and thus we cannot query it
-          //   continue
-          // }
-          //
-          // if (!metaData[key] || !metaData[key].meta || !metaData[key].meta[branchKey]) {
-          //   continue
-          // }
+        // if (!hits.find((h) => h !== 0)) {
+        //   // if there are no hits the meta object is not created and thus we cannot query it
+        //   continue
+        // }
+        //
+        // if (!metaData[key] || !metaData[key].meta || !metaData[key].meta[branchKey]) {
+        //   continue
+        // }
 
-          const meta = metaData?.[key]?.meta?.[branchKey]
+        const meta = metaData?.[key]?.meta?.[branchKey];
 
-          traces.push({
-            id: `b-${branch.line}`,
-            path: key,
-            type: "branch",
-            line: branch.line,
+        traces.push({
+          id: `b-${branch.line}`,
+          path: key,
+          type: "branch",
+          line: branch.line,
 
-            locationIdx: 0,
-            branchType: true,
+          locationIdx: 0,
+          branchType: true,
 
-            hits: hits[0],
+          hits: hits[0],
 
-            condition_ast: meta?.condition_ast,
-            condition: meta?.condition,
-            variables: meta?.variables
-          });
+          condition_ast: meta?.condition_ast,
+          condition: meta?.condition,
+          variables: meta?.variables,
+        });
 
-          traces.push({
-            id: `b-${branch.line}`,
-            path: key,
-            type: "branch",
-            line: branch.line,
+        traces.push({
+          id: `b-${branch.line}`,
+          path: key,
+          type: "branch",
+          line: branch.line,
 
-            locationIdx: 1,
-            branchType: false,
+          locationIdx: 1,
+          branchType: false,
 
-            hits: hits[1],
+          hits: hits[1],
 
-            condition_ast: meta?.condition_ast,
-            condition: meta?.condition,
-            variables: meta?.variables
-          });
-        }
+          condition_ast: meta?.condition_ast,
+          condition: meta?.condition,
+          variables: meta?.variables,
+        });
+      }
     }
 
     // Retrieve execution information
@@ -297,16 +343,21 @@ export class JavaScriptRunner implements EncodingRunner<JavaScriptTestCase> {
     return executionResult;
   }
 
+<<<<<<< HEAD
   resetInstrumentationData () {
     for (const key of Object.keys(global.__coverage__ || {})) {
+=======
+  resetInstrumentationData() {
+    for (const key of Object.keys(global.__coverage__)) {
+>>>>>>> develop
       for (const statementKey of Object.keys(global.__coverage__[key].s)) {
-        global.__coverage__[key].s[statementKey] = 0
+        global.__coverage__[key].s[statementKey] = 0;
       }
       for (const functionKey of Object.keys(global.__coverage__[key].f)) {
-        global.__coverage__[key].f[functionKey] = 0
+        global.__coverage__[key].f[functionKey] = 0;
       }
       for (const branchKey of Object.keys(global.__coverage__[key].b)) {
-        global.__coverage__[key].b[branchKey] = [0, 0]
+        global.__coverage__[key].b[branchKey] = [0, 0];
       }
     }
   }
