@@ -1,4 +1,4 @@
-const Mocha = require('mocha')
+import Mocha = require("mocha");
 
 const {
   EVENT_RUN_BEGIN,
@@ -6,17 +6,16 @@ const {
   EVENT_TEST_FAIL,
   EVENT_TEST_PASS,
   EVENT_SUITE_BEGIN,
-  EVENT_SUITE_END
+  EVENT_SUITE_END,
 } = Mocha.Runner.constants;
 
 export class SilentMochaReporter {
-  private _indents: number
+  private _indents: number;
   private failures;
 
   constructor(runner) {
     this._indents = 0;
-    this.failures = []
-    const stats = runner.stats;
+    this.failures = [];
 
     runner
       .once(EVENT_RUN_BEGIN, () => {
@@ -28,17 +27,17 @@ export class SilentMochaReporter {
       .on(EVENT_SUITE_END, () => {
         this.decreaseIndent();
       })
-      .on(EVENT_TEST_PASS, test => {
+      .on(EVENT_TEST_PASS, (test) => {
         // Test#fullTitle() returns the suite name(s)
         // prepended to the test title
         console.log(`${this.indent()}pass: ${test.fullTitle()}`);
 
         if (test.duration > test.slow()) {
-          test.speed = 'slow';
+          test.speed = "slow";
         } else if (test.duration > test.slow() / 2) {
-          test.speed = 'medium';
+          test.speed = "medium";
         } else {
-          test.speed = 'fast';
+          test.speed = "fast";
         }
       })
       .on(EVENT_TEST_FAIL, (test, err) => {
@@ -64,7 +63,7 @@ export class SilentMochaReporter {
   }
 
   indent() {
-    return Array(this._indents).join('  ');
+    return Array(this._indents).join("  ");
   }
 
   increaseIndent() {
