@@ -18,7 +18,7 @@
 import * as path from "path";
 import { getAllFiles, readFile } from "../../utils/fileSystem";
 import { AbstractSyntaxTreeGenerator } from "./ast/AbstractSyntaxTreeGenerator";
-import { CFG, Properties, TargetMetaData, TargetPool } from "@syntest/framework";
+import { CFG, Properties, TargetMetaData, TargetPool } from "@syntest/core";
 import { TargetMapGenerator } from "./map/TargetMapGenerator";
 import { ControlFlowGraphGenerator } from "./cfg/ControlFlowGraphGenerator";
 import { ImportGenerator } from "./dependency/ImportGenerator";
@@ -37,13 +37,13 @@ import { TypeEnum } from "./types/resolving/TypeEnum";
 import { TypeProbability } from "./types/resolving/TypeProbability";
 import { Instrumenter } from "../../instrumentation/Instrumenter";
 import { ExportType } from "./dependency/IdentifierVisitor";
-
+import * as t from "@babel/types";
+// eslint-disable-next-line
 const { outputFileSync, copySync } = require("fs-extra");
 
-
 export interface JavaScriptTargetMetaData extends TargetMetaData {
-  type: SubjectType,
-  export: Export
+  type: SubjectType;
+  export: Export;
 }
 
 export class JavaScriptTargetPool extends TargetPool {
@@ -58,7 +58,7 @@ export class JavaScriptTargetPool extends TargetPool {
   protected _sources: Map<string, string>;
 
   // Mapping: filepath -> AST
-  protected _abstractSyntaxTrees: Map<string, any>;
+  protected _abstractSyntaxTrees: Map<string, t.Node>;
 
   // Mapping: filepath -> target name -> target meta data
   protected _targetMap: Map<string, Map<string, JavaScriptTargetMetaData>>;
@@ -76,7 +76,7 @@ export class JavaScriptTargetPool extends TargetPool {
   protected _dependencyMaps: Map<string, Export[]>;
 
   // Mapping: filepath -> target name -> Exports
-  protected _exportMap: Map<string, Export[]>
+  protected _exportMap: Map<string, Export[]>;
 
   constructor(
     abstractSyntaxTreeGenerator: AbstractSyntaxTreeGenerator,
@@ -95,7 +95,7 @@ export class JavaScriptTargetPool extends TargetPool {
     this._typeResolver = typeResolver;
 
     this._sources = new Map<string, string>();
-    this._abstractSyntaxTrees = new Map<string, string>();
+    this._abstractSyntaxTrees = new Map<string, t.Node>();
     this._targetMap = new Map<string, Map<string, JavaScriptTargetMetaData>>();
     this._functionMaps = new Map<
       string,
@@ -105,7 +105,7 @@ export class JavaScriptTargetPool extends TargetPool {
 
     this._dependencyMaps = new Map();
 
-    this._exportMap = new Map()
+    this._exportMap = new Map();
   }
 
   getSource(targetPath: string) {
@@ -113,37 +113,36 @@ export class JavaScriptTargetPool extends TargetPool {
 
     if (!this._sources.has(absoluteTargetPath)) {
       if (!existsSync(absoluteTargetPath)) {
-        if (existsSync(absoluteTargetPath + '.js')) {
-          absoluteTargetPath += '.js'
-        } else if (existsSync(absoluteTargetPath + '.ts')) {
-          absoluteTargetPath += '.ts'
+        if (existsSync(absoluteTargetPath + ".js")) {
+          absoluteTargetPath += ".js";
+        } else if (existsSync(absoluteTargetPath + ".ts")) {
+          absoluteTargetPath += ".ts";
         } else {
-          throw new Error("Cannot find source: " + absoluteTargetPath)
+          throw new Error("Cannot find source: " + absoluteTargetPath);
         }
       }
 
-      const stats = lstatSync(absoluteTargetPath)
+      const stats = lstatSync(absoluteTargetPath);
 
       if (stats.isDirectory()) {
-        if (existsSync(absoluteTargetPath + '/index.js')) {
-          absoluteTargetPath += '/index.js'
-        } else if (existsSync(absoluteTargetPath + '/index.ts')) {
-          absoluteTargetPath += '/index.ts'
+        if (existsSync(absoluteTargetPath + "/index.js")) {
+          absoluteTargetPath += "/index.js";
+        } else if (existsSync(absoluteTargetPath + "/index.ts")) {
+          absoluteTargetPath += "/index.ts";
         } else {
-          throw new Error("Cannot find source: " + absoluteTargetPath)
+          throw new Error("Cannot find source: " + absoluteTargetPath);
         }
       }
 
-      return readFile(absoluteTargetPath)
+      return readFile(absoluteTargetPath);
       // this._sources.set(absoluteTargetPath, readFile(absoluteTargetPath));
     }
 
     return this._sources.get(absoluteTargetPath);
   }
 
-  getAST(targetPath: string) {
+  getAST(targetPath: string): t.Node {
     const absoluteTargetPath = path.resolve(targetPath);
-
 
     // this takes up too much memory we should do some kind of garbage collection if we want to save it all
     if (!this._abstractSyntaxTrees.has(absoluteTargetPath)) {
@@ -157,7 +156,7 @@ export class JavaScriptTargetPool extends TargetPool {
       return this.abstractSyntaxTreeGenerator.generate(
         this.getSource(targetPath),
         absoluteTargetPath
-      )
+      );
     }
 
     return this._abstractSyntaxTrees.get(absoluteTargetPath);
@@ -167,7 +166,7 @@ export class JavaScriptTargetPool extends TargetPool {
     const absoluteTargetPath = path.resolve(targetPath);
 
     if (!this._controlFlowGraphs.has(absoluteTargetPath)) {
-      this._controlFlowGraphs.set(absoluteTargetPath, new Map<string, CFG>())
+      this._controlFlowGraphs.set(absoluteTargetPath, new Map<string, CFG>());
     }
 
     if (!this._controlFlowGraphs.get(absoluteTargetPath).has(targetName)) {
@@ -179,7 +178,7 @@ export class JavaScriptTargetPool extends TargetPool {
       // );
       return this.controlFlowGraphGenerator.convertAST(
         this.getAST(absoluteTargetPath)
-      )
+      );
     }
 
     return this._controlFlowGraphs.get(absoluteTargetPath).get(targetName);
@@ -190,31 +189,38 @@ export class JavaScriptTargetPool extends TargetPool {
 
     if (!this._targetMap.has(absoluteTargetPath)) {
       const targetAST = this.getAST(absoluteTargetPath);
-      const { targetMap, functionMap } =
-        this.targetMapGenerator.generate(absoluteTargetPath, targetAST);
+      const { targetMap, functionMap } = this.targetMapGenerator.generate(
+        absoluteTargetPath,
+        targetAST
+      );
 
-      const exports = this.getExports(targetPath)
+      const exports = this.getExports(targetPath);
 
-      const finalTargetMap = new Map<string, JavaScriptTargetMetaData>()
+      const finalTargetMap = new Map<string, JavaScriptTargetMetaData>();
 
       for (const key of targetMap.keys()) {
-        const name = targetMap.get(key).name
-        const export_ = exports.find((e) => e.name === name)
+        const name = targetMap.get(key).name;
+        const export_ = exports.find((e) => e.name === name);
 
         if (!export_) {
           // No export found so we cannot import it and thus not test it
-          continue
+          continue;
         }
 
-        if(export_.type === ExportType.const && functionMap.get(key).size === 0) {
-          throw new Error(`Target cannot be constant: ${name} -> ${JSON.stringify(export_)}`)
+        if (
+          export_.type === ExportType.const &&
+          functionMap.get(key).size === 0
+        ) {
+          throw new Error(
+            `Target cannot be constant: ${name} -> ${JSON.stringify(export_)}`
+          );
         }
 
-        let isPrototypeClass = false
+        let isPrototypeClass = false;
         for (const func of functionMap.get(key).values()) {
           if (func.isConstructor) {
-            isPrototypeClass = true
-            break
+            isPrototypeClass = true;
+            break;
           }
         }
 
@@ -226,11 +232,14 @@ export class JavaScriptTargetPool extends TargetPool {
         // threat everything as a function if we don't know
         finalTargetMap.set(key, {
           name: name,
-          type: export_.type === ExportType.class || isPrototypeClass
-            ? SubjectType.class
-            : (export_.type === ExportType.const ? SubjectType.object : SubjectType.function),
-          export: export_
-        })
+          type:
+            export_.type === ExportType.class || isPrototypeClass
+              ? SubjectType.class
+              : export_.type === ExportType.const
+              ? SubjectType.object
+              : SubjectType.function,
+          export: export_,
+        });
       }
 
       this._targetMap.set(absoluteTargetPath, finalTargetMap);
@@ -240,14 +249,14 @@ export class JavaScriptTargetPool extends TargetPool {
     return this._targetMap.get(absoluteTargetPath);
   }
 
-  getFunctionMap(
+  getFunctionMapSpecific(
     targetPath: string,
     targetName: string
   ): Map<string, ActionDescription> {
     const absoluteTargetPath = path.resolve(targetPath);
 
     if (!this._functionMaps.has(absoluteTargetPath)) {
-      this.getTargetMap(absoluteTargetPath)
+      this.getTargetMap(absoluteTargetPath);
     }
 
     if (this._functionMaps.get(absoluteTargetPath).has(targetName)) {
@@ -259,25 +268,43 @@ export class JavaScriptTargetPool extends TargetPool {
     }
   }
 
+  getFunctionMap(
+    targetPath: string
+  ): Map<string, Map<string, ActionDescription>> {
+    const absoluteTargetPath = path.resolve(targetPath);
+
+    if (!this._functionMaps.has(absoluteTargetPath)) {
+      this.getTargetMap(absoluteTargetPath);
+    }
+
+    return this._functionMaps.get(absoluteTargetPath);
+  }
+
   getExports(targetPath: string): Export[] {
     const absoluteTargetPath = path.resolve(targetPath);
 
     if (!this._exportMap.has(absoluteTargetPath)) {
-      const exports = this.exportGenerator.generate(absoluteTargetPath, this.getAST(absoluteTargetPath))
+      const exports = this.exportGenerator.generate(
+        absoluteTargetPath,
+        this.getAST(absoluteTargetPath)
+      );
 
       // this._exportMap.set(absoluteTargetPath, exports);
-      return exports
+      return exports;
     }
 
-    return this._exportMap.get(absoluteTargetPath)
+    return this._exportMap.get(absoluteTargetPath);
   }
 
-  getDependencies(targetPath: string):  Export[] {
+  getDependencies(targetPath: string): Export[] {
     const absoluteTargetPath = path.resolve(targetPath);
 
     if (!this._dependencyMaps.has(absoluteTargetPath)) {
       // Find all external imports in the file under test
-      const imports = this.importGenerator.generate(absoluteTargetPath, this.getAST(targetPath))
+      const imports = this.importGenerator.generate(
+        absoluteTargetPath,
+        this.getAST(targetPath)
+      );
 
       // For each external import scan the file for libraries with exported functions
       const libraries: Export[] = [];
@@ -287,17 +314,16 @@ export class JavaScriptTargetPool extends TargetPool {
 
         // External libraries have a different path!
         try {
-          this.getSource(pathLib)
+          this.getSource(pathLib);
         } catch (e) {
-          if (e.message.includes('Cannot find source')) {
+          if (e.message.includes("Cannot find source")) {
             // TODO would be nice if we could get the actual path! (node modules)
-            return
+            return;
             // pathLib = path.join
           } else {
-            throw e
+            throw e;
           }
         }
-
 
         // Scan for libraries with public or external functions
         let exports = [];
@@ -315,29 +341,32 @@ export class JavaScriptTargetPool extends TargetPool {
         }
       });
 
-      return libraries
+      return libraries;
       // this._dependencyMaps.set(targetPath, libraries);
     }
 
-    return this._dependencyMaps.get(absoluteTargetPath)
+    return this._dependencyMaps.get(absoluteTargetPath);
   }
 
   async prepareAndInstrument(): Promise<void> {
-    const absoluteRootPath = path.resolve(Properties.target_root_directory)
+    const absoluteRootPath = path.resolve(Properties.target_root_directory);
 
-    const destinationPath = path.resolve(Properties.temp_instrumented_directory, path.basename(Properties.target_root_directory))
+    const destinationPath = path.resolve(
+      Properties.temp_instrumented_directory,
+      path.basename(Properties.target_root_directory)
+    );
 
     console.log(absoluteRootPath, destinationPath);
     // copy everything
-    await copySync(absoluteRootPath, destinationPath)
+    await copySync(absoluteRootPath, destinationPath);
 
     // overwrite the stuff that needs instrumentation
     const instrumenter = new Instrumenter();
 
-    const targetPaths = this.targets.map((x) => x.canonicalPath)
+    const targetPaths = this.targets.map((x) => x.canonicalPath);
 
     for (const targetPath of targetPaths) {
-      const source = this.getSource(targetPath)
+      const source = this.getSource(targetPath);
       const instrumentedSource = await instrumenter.instrument(
         source,
         targetPath
@@ -345,71 +374,139 @@ export class JavaScriptTargetPool extends TargetPool {
 
       const _path = path
         .normalize(targetPath)
-        .replace(
-          absoluteRootPath,
-          destinationPath
-        )
+        .replace(absoluteRootPath, destinationPath);
 
       await outputFileSync(_path, instrumentedSource);
     }
   }
 
   scanTargetRootDirectory(): void {
-    const absoluteRootPath = path.resolve(Properties.target_root_directory)
+    const absoluteRootPath = path.resolve(Properties.target_root_directory);
 
     // TODO remove the filters
-    const files = getAllFiles(absoluteRootPath, ".js")
-      .filter((x) => !x.includes('/test/')
-        && !x.includes('.test.js')
-        && !x.includes('node_modules')) // maybe we should also take those into account
+    const files = getAllFiles(absoluteRootPath, ".js").filter(
+      (x) =>
+        !x.includes("/test/") &&
+        !x.includes(".test.js") &&
+        !x.includes("node_modules")
+    ); // maybe we should also take those into account
 
-    const objects: ComplexObject[] = []
-    const objectGenerator = new ObjectGenerator()
+    const objects: ComplexObject[] = [];
+    const objectGenerator = new ObjectGenerator();
 
     for (const file of files) {
-      const exports = this.getExports(file)
-      objects.push(...objectGenerator.generate(file, this.getAST(file), exports))
+      const exports = this.getExports(file);
+      objects.push(
+        ...objectGenerator.generate(file, this.getAST(file), exports)
+      );
     }
 
     // standard stuff
     // function https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function
     objects.push({
       name: "function",
-      properties: new Set(['arguments', 'caller', 'displayName', 'length', 'name']),
-      functions: new Set(['apply', 'bind', 'call', 'toString']),
+      properties: new Set([
+        "arguments",
+        "caller",
+        "displayName",
+        "length",
+        "name",
+      ]),
+      functions: new Set(["apply", "bind", "call", "toString"]),
       propertyType: new Map<string, TypeProbability>([
-        ['arguments', new TypeProbability([[TypeEnum.ARRAY, 1, null]])],
-        ['caller', new TypeProbability([[TypeEnum.FUNCTION, 1, null]])],
-        ['displayName', new TypeProbability([[TypeEnum.STRING, 1, null]])],
-        ['length', new TypeProbability([[TypeEnum.NUMERIC, 1, null]])],
-        ['name', new TypeProbability([[TypeEnum.STRING, 1, null]])]
-      ])
-    })
+        ["arguments", new TypeProbability([[TypeEnum.ARRAY, 1, null]])],
+        ["caller", new TypeProbability([[TypeEnum.FUNCTION, 1, null]])],
+        ["displayName", new TypeProbability([[TypeEnum.STRING, 1, null]])],
+        ["length", new TypeProbability([[TypeEnum.NUMERIC, 1, null]])],
+        ["name", new TypeProbability([[TypeEnum.STRING, 1, null]])],
+      ]),
+    });
 
     // array https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array
     objects.push({
       name: "array",
-      properties: new Set(['length']),
-      functions: new Set(['at', 'concat', 'copyWithin', 'entries', 'fill', 'filter', 'find', 'findIndex', 'flat', 'flatMap', 'includes', 'indexOf', 'join', 'keys', 'lastIndexOf', 'map', 'pop', 'push', 'reduce', 'reduceRight', 'reverse', 'shift', 'slice', 'toLocaleString', 'toString', 'unshift', 'values']),
+      properties: new Set(["length"]),
+      functions: new Set([
+        "at",
+        "concat",
+        "copyWithin",
+        "entries",
+        "fill",
+        "filter",
+        "find",
+        "findIndex",
+        "flat",
+        "flatMap",
+        "includes",
+        "indexOf",
+        "join",
+        "keys",
+        "lastIndexOf",
+        "map",
+        "pop",
+        "push",
+        "reduce",
+        "reduceRight",
+        "reverse",
+        "shift",
+        "slice",
+        "toLocaleString",
+        "toString",
+        "unshift",
+        "values",
+      ]),
       propertyType: new Map<string, TypeProbability>([
-        ['length', new TypeProbability([[TypeEnum.NUMERIC, 1, null]])]
-      ])
-    })
+        ["length", new TypeProbability([[TypeEnum.NUMERIC, 1, null]])],
+      ]),
+    });
 
     // string
     objects.push({
       name: "string",
-      properties: new Set(['length']),
-      functions: new Set(['at', 'charAt', 'charCodeAt', 'codePointAt', 'concat', 'includes', 'endsWith', 'indexOf', 'lastIndexOf', 'localeCompare', 'match', 'matchAll', 'normalize', 'padEnd', 'padStart', 'repeat', 'replace', 'replaceAll', 'search', 'slice', 'split', 'startsWith', 'substring', 'toLocaleLowerCase', 'toLocaleUpperCase', 'toLowerCase', 'toString', 'toUpperCase', 'trim', 'trimStart', 'trimEnd', 'valueOf']),
+      properties: new Set(["length"]),
+      functions: new Set([
+        "at",
+        "charAt",
+        "charCodeAt",
+        "codePointAt",
+        "concat",
+        "includes",
+        "endsWith",
+        "indexOf",
+        "lastIndexOf",
+        "localeCompare",
+        "match",
+        "matchAll",
+        "normalize",
+        "padEnd",
+        "padStart",
+        "repeat",
+        "replace",
+        "replaceAll",
+        "search",
+        "slice",
+        "split",
+        "startsWith",
+        "substring",
+        "toLocaleLowerCase",
+        "toLocaleUpperCase",
+        "toLowerCase",
+        "toString",
+        "toUpperCase",
+        "trim",
+        "trimStart",
+        "trimEnd",
+        "valueOf",
+      ]),
       propertyType: new Map<string, TypeProbability>([
-        ['length', new TypeProbability([[TypeEnum.NUMERIC, 1, null]])]
-      ])
-    })
+        ["length", new TypeProbability([[TypeEnum.NUMERIC, 1, null]])],
+      ]),
+    });
 
     // TODO npm dependencies
     // TODO get rid of duplicates
 
-    const finalObjects = []
+    const finalObjects = [];
 
     function eqSet(as, bs) {
       if (as.size !== bs.size) return false;
@@ -419,39 +516,46 @@ export class JavaScriptTargetPool extends TargetPool {
 
     objects.forEach((o) => {
       if (!o.properties.size && !o.functions.size) {
-        return
+        return;
       }
 
       const found = finalObjects.find((o2) => {
-        return o.export === o2.export // TODO not sure if you can compare exports like this
-          && o.name === o2.name
-          && eqSet(o.properties, o2.properties)
-          && eqSet(o.functions, o2.functions)
-      })
+        return (
+          o.export === o2.export && // TODO not sure if you can compare exports like this
+          o.name === o2.name &&
+          eqSet(o.properties, o2.properties) &&
+          eqSet(o.functions, o2.functions)
+        );
+      });
 
       if (!found) {
-        finalObjects.push(o)
+        finalObjects.push(o);
       }
-    })
+    });
 
-
-    const generator = new VariableGenerator()
-    const elements: Element[] = []
-    const relations: Relation[] = []
-    const wrapperElementIsRelation: Map<string, Relation> = new Map()
+    const generator = new VariableGenerator();
+    const elements: Element[] = [];
+    const relations: Relation[] = [];
+    const wrapperElementIsRelation: Map<string, Relation> = new Map();
 
     for (const file of files) {
-      const [_elements, _relations, _wrapperElementIsRelation] = generator.generate(file, this.getAST(file))
+      const [_elements, _relations, _wrapperElementIsRelation] =
+        generator.generate(file, this.getAST(file));
 
-      elements.push(..._elements)
-      relations.push(..._relations)
+      elements.push(..._elements);
+      relations.push(..._relations);
 
       for (const key of _wrapperElementIsRelation.keys()) {
-        wrapperElementIsRelation.set(key, _wrapperElementIsRelation.get(key))
+        wrapperElementIsRelation.set(key, _wrapperElementIsRelation.get(key));
       }
     }
 
-    this._typeResolver.resolveTypes(elements, relations, wrapperElementIsRelation, finalObjects)
+    this._typeResolver.resolveTypes(
+      elements,
+      relations,
+      wrapperElementIsRelation,
+      finalObjects
+    );
   }
 
   get typeResolver(): TypeResolver {
