@@ -18,12 +18,16 @@
 import {
   BranchNode,
   CFG,
-  CFGFactory, Edge,
+  CFGFactory,
+  Edge,
   Node,
-  NodeType, Operation,
-  PlaceholderNode, prng,
+  NodeType,
+  Operation,
+  PlaceholderNode,
+  prng,
   RootNode,
-} from "@syntest/framework";
+} from "@syntest/core";
+import * as t from "@babel/types";
 
 interface ReturnValue {
   childNodes: Node[];
@@ -31,14 +35,10 @@ interface ReturnValue {
 }
 
 export class ControlFlowGraphGenerator implements CFGFactory {
-
   private cfg: CFG;
-  private _contracts: string[] = [];
 
-  convertAST(ast: any, compress = false, placeholder = false): CFG {
+  convertAST(ast: t.Node): CFG {
     // TODO the imported stuff should also be resolved...
-    this._contracts = [];
-
     this.cfg = new CFG();
 
     this.visitChild(ast, []);
@@ -53,7 +53,6 @@ export class ControlFlowGraphGenerator implements CFGFactory {
 
     return this.cfg;
   }
-
 
   removePlaceholder(): void {
     const removableEdges = [];
@@ -212,7 +211,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
       lines: lines,
       statements: statements,
       type: NodeType.Root,
-      description: description
+      description: description,
     };
 
     this.cfg.nodes.push(node);
@@ -242,7 +241,6 @@ export class ControlFlowGraphGenerator implements CFGFactory {
 
     return node;
   }
-
 
   private createPlaceholderNode(
     lines: number[],
@@ -302,14 +300,9 @@ export class ControlFlowGraphGenerator implements CFGFactory {
    *
    * @param child the child ast node
    * @param parents the parents of the child
-   * @param contractName
    * @private
    */
-  private visitChild(
-    child: any,
-    parents: Node[],
-    contractName?: string
-  ): ReturnValue {
+  private visitChild(child: t.Node, parents: Node[]): ReturnValue {
     const skipable: string[] = [
       "ImportDeclaration",
       "ClassProperty",
@@ -332,9 +325,9 @@ export class ControlFlowGraphGenerator implements CFGFactory {
       //     breakNodes: []
       //   }
       case "File":
-        return this.visitChild(child.program, parents)
+        return this.visitChild(child.program, parents);
       case "ExportDefaultDeclaration":
-        return this.visitChild(child.declaration, parents)
+        return this.visitChild(child.declaration, parents);
       case "ExportSpecifier":
         return this.visitChild(child.local, parents);
       case "UnaryExpression":
@@ -344,9 +337,9 @@ export class ControlFlowGraphGenerator implements CFGFactory {
 
       //
       case "Program":
-        return this.visitProgram(child)
+        return this.visitProgram(child);
       case "FunctionDeclaration":
-        return this.visitFunctionDeclaration(child)
+        return this.visitFunctionDeclaration(child);
       case "CallExpression":
         return this.visitCallExpression(child, parents);
       case "ExportNamedDeclaration":
@@ -383,20 +376,20 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         return this.visitGeneralExpression(child, parents);
 
       case "VariableDeclaration":
-        return this.visitVariableDeclaration(child, parents)
+        return this.visitVariableDeclaration(child, parents);
 
       case "ReturnStatement":
-        return this.visitReturnStatement(child, parents)
+        return this.visitReturnStatement(child, parents);
       case "ThrowStatement":
-        return this.visitThrowStatement(child, parents)
+        return this.visitThrowStatement(child, parents);
 
       case "BlockStatement":
-        return this.visitBlockStatement(child, parents)
+        return this.visitBlockStatement(child, parents);
 
       case "ClassDeclaration":
-        return this.visitClassDeclaration(child)
+        return this.visitClassDeclaration(child);
       case "ClassMethod":
-        return this.visitClassMethod(child)
+        return this.visitClassMethod(child);
 
       case "IfStatement":
         return this.visitIfStatement(child, parents);
@@ -404,7 +397,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         return this.visitConditional(child, parents);
 
       case "TryStatement":
-        return this.visitTryStatement(child, parents)
+        return this.visitTryStatement(child, parents);
 
       case "CatchClause":
         return this.visitCatchClause(child, parents);
@@ -414,23 +407,21 @@ export class ControlFlowGraphGenerator implements CFGFactory {
       case "DoWhileStatement":
         return this.visitDoWhileStatement(child, parents);
       case "ForOfStatement":
-        return this.visitForOfStatement(child, parents)
+        return this.visitForOfStatement(child, parents);
       case "ForInStatement":
-        return this.visitForOfStatement(child, parents)
+        return this.visitForOfStatement(child, parents);
       case "ForStatement":
         return this.visitForStatement(child, parents);
 
       case "BreakStatement":
-        return this.visitBreakStatement(child, parents)
+        return this.visitBreakStatement(child, parents);
       case "ContinueStatement":
-        return this.visitContinueStatement(child, parents)
-
+        return this.visitContinueStatement(child, parents);
 
       case "SwitchStatement":
-        return this.visitSwitchStatement(child, parents)
+        return this.visitSwitchStatement(child, parents);
       case "SwitchCase":
-        return this.visitSwitchCase(child, parents)
-
+        return this.visitSwitchCase(child, parents);
 
       // case "SourceUnit":
       //   return this.SourceUnit(cfg, child);
@@ -447,7 +438,6 @@ export class ControlFlowGraphGenerator implements CFGFactory {
       //
       // case "IfStatement":
       //   return this.IfStatement(cfg, child, parents);
-
 
       // case "WhileStatement":
       //   return this.WhileStatement(cfg, child, parents);
@@ -469,7 +459,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     }
   }
 
-  private visitProgram(ast: any): ReturnValue {
+  private visitProgram(ast: t.Program): ReturnValue {
     for (const child of ast.body) {
       // TODO add more probably
       // if (!['FunctionDeclaration', 'ClassDeclaration', 'ExpressionStatement'].includes(child?.type)) {
@@ -486,7 +476,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
-  private visitFunctionDeclaration(ast: any): ReturnValue {
+  private visitFunctionDeclaration(ast: t.FunctionDeclaration): ReturnValue {
     if (!ast.loc) {
       return {
         childNodes: [],
@@ -499,10 +489,9 @@ export class ControlFlowGraphGenerator implements CFGFactory {
       ast.id.name
     );
 
+    const parents: Node[] = [node];
 
-    let parents: Node[] = [node];
-
-    const totalBreakNodes = [];
+    // const totalBreakNodes = [];
     // if (ast.modifiers && Properties.modifier_extraction) {
     //   ast.modifiers.forEach((modifier) => {
     //     const { childNodes, breakNodes } = this.visitChild(
@@ -530,50 +519,54 @@ export class ControlFlowGraphGenerator implements CFGFactory {
   }
 
   private visitCallExpression(
-    ast: any,
+    ast: t.CallExpression,
     parents: Node[]
   ): ReturnValue {
     const node: Node = this.createNode([ast.loc?.start?.line], []);
     this.connectParents(parents, [node]);
-    let nodes = [node]
+    let nodes = [node];
     for (const arg of ast.arguments) {
       const result = this.visitChild(arg, nodes);
 
-      nodes = result.childNodes
+      nodes = result.childNodes;
     }
 
     return {
       childNodes: [node],
-      breakNodes: []
-    }
+      breakNodes: [],
+    };
   }
 
   private visitExportNamedDeclaration(
-    ast: any,
+    ast: t.ExportNamedDeclaration,
     parents: Node[]
   ): ReturnValue {
     if (ast.specifiers && ast.specifiers.length) {
-      let nodes = parents
+      let nodes = parents;
       for (const specifier of ast.specifiers) {
         const result = this.visitChild(specifier, nodes);
 
-        nodes = result.childNodes
+        nodes = result.childNodes;
       }
 
       return {
         childNodes: nodes,
-        breakNodes: []
-      }
+        breakNodes: [],
+      };
     } else {
-      return this.visitChild(ast.declaration, parents)
+      return this.visitChild(ast.declaration, parents);
     }
   }
 
   private visitGeneralExpression(
-    ast: any,
+    ast: t.Expression | t.VariableDeclarator | t.SpreadElement,
     parents: Node[]
   ): ReturnValue {
+<<<<<<< HEAD
     if (['LogicalExpression', 'BinaryExpression'].includes(ast?.type)) {
+=======
+    if (ast.type === "BinaryExpression" || ast.type === "LogicalExpression") {
+>>>>>>> develop
       const left = this.visitChild(ast.left, parents);
       const right = this.visitChild(ast.right, left.childNodes);
 
@@ -581,9 +574,15 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         childNodes: [...right.childNodes],
         breakNodes: [...left.breakNodes, ...right.breakNodes],
       };
+<<<<<<< HEAD
     } else if (ast?.type === 'AssignmentExpression') {
       return this.visitChild(ast.right, parents);
     } else if (ast?.type === 'VariableDeclarator') {
+=======
+    } else if (ast.type === "AssignmentExpression") {
+      return this.visitChild(ast.right, parents);
+    } else if (ast.type === "VariableDeclarator") {
+>>>>>>> develop
       if (ast.init) {
         return this.visitChild(ast.init, parents);
       }
@@ -591,17 +590,24 @@ export class ControlFlowGraphGenerator implements CFGFactory {
       const node: Node = this.createNode([ast.loc?.start?.line], []);
       this.connectParents(parents, [node]);
 
-
       return {
         childNodes: [node],
         breakNodes: [],
       };
+<<<<<<< HEAD
     } else if (ast?.type === 'FunctionExpression'
     || ast?.type === 'ArrowFunctionExpression') {
       const node: RootNode = this.createRootNode(
         [ast.loc?.start.line],
         [],
       );
+=======
+    } else if (
+      ast.type === "FunctionExpression" ||
+      ast.type === "ArrowFunctionExpression"
+    ) {
+      const node: RootNode = this.createRootNode([ast.loc.start.line], []);
+>>>>>>> develop
       this.connectParents(parents, [node]);
 
       if (ast.body) {
@@ -621,10 +627,12 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         breakNodes: [],
       };
     }
-
   }
 
-  private visitBlockStatement(ast: any, parents: Node[]): ReturnValue {
+  private visitBlockStatement(
+    ast: t.BlockStatement,
+    parents: Node[]
+  ): ReturnValue {
     let nodes = parents;
     const totalBreakNodes = [];
     for (const child of ast?.body) {
@@ -646,11 +654,14 @@ export class ControlFlowGraphGenerator implements CFGFactory {
    * @constructor
    * @private
    */
-  private visitReturnStatement(ast: any, parents: Node[]): ReturnValue {
-    const count = this.cfg.edges.length
+  private visitReturnStatement(
+    ast: t.ReturnStatement,
+    parents: Node[]
+  ): ReturnValue {
+    const count = this.cfg.edges.length;
 
     if (ast.argument) {
-      this.visitChild(ast.argument, parents)
+      this.visitChild(ast.argument, parents);
     }
 
     if (!this.cfg.edges[count]) {
@@ -666,8 +677,16 @@ export class ControlFlowGraphGenerator implements CFGFactory {
   }
 
   // TODO try catch support
+<<<<<<< HEAD
   private visitThrowStatement(ast: any, parents: Node[]): ReturnValue {
     const node: Node = this.createNode([ast.loc?.start.line], []);
+=======
+  private visitThrowStatement(
+    ast: t.ThrowStatement,
+    parents: Node[]
+  ): ReturnValue {
+    const node: Node = this.createNode([ast.loc.start.line], []);
+>>>>>>> develop
     this.connectParents(parents, [node]);
 
     return {
@@ -676,8 +695,8 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
-  private visitClassDeclaration(ast: any): ReturnValue {
-    const classBody = ast.body
+  private visitClassDeclaration(ast: t.ClassDeclaration): ReturnValue {
+    const classBody = ast.body;
     for (const child of classBody.body) {
       this.visitChild(child, []);
     }
@@ -688,14 +707,14 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
-  private visitClassMethod(ast: any): ReturnValue {
+  private visitClassMethod(ast: t.ClassMethod): ReturnValue {
     const node: RootNode = this.createRootNode(
       [ast.loc?.start.line],
       [],
-      ast.key.name
+      ast.key.type === "Identifier" ? ast.key.name : ""
     );
 
-    let parents: Node[] = [node];
+    const parents: Node[] = [node];
 
     // Check if body is block
     if (ast.body) {
@@ -709,13 +728,13 @@ export class ControlFlowGraphGenerator implements CFGFactory {
   }
 
   private visitVariableDeclaration(
-    ast: any,
+    ast: t.VariableDeclaration,
     parents: Node[]
   ): ReturnValue {
     // const node: Node = this.createNode([ast.loc?.start.line], []);
     // this.connectParents(parents, [node]);
 
-    let nodes = parents
+    let nodes = parents;
     const totalBreakNodes = [];
     for (const child of ast.declarations) {
       const { childNodes, breakNodes } = this.visitChild(child, nodes);
@@ -729,6 +748,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
+<<<<<<< HEAD
   private visitIfStatement(ast: any, parents: Node[]): ReturnValue {
     const node: BranchNode = this.createBranchNode(
       [ast.loc?.start.line],
@@ -738,6 +758,17 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         operator: ast.test.operator,
       }
     );
+=======
+  private visitIfStatement(ast: t.IfStatement, parents: Node[]): ReturnValue {
+    const node: BranchNode = this.createBranchNode([ast.loc.start.line], [], {
+      type: ast.test.type,
+      operator:
+        ast.test.type === "BinaryExpression" ||
+        ast.test.type === "UnaryExpression"
+          ? ast.test.operator
+          : "",
+    });
+>>>>>>> develop
 
     this.connectParents(parents, [node]);
 
@@ -746,9 +777,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
 
     // Visit true flow
     let count = this.cfg.edges.length;
-    const { childNodes, breakNodes } = this.visitChild(ast.consequent, [
-      node,
-    ]);
+    const { childNodes, breakNodes } = this.visitChild(ast.consequent, [node]);
     const trueNodes = childNodes;
 
     totalBreakNodes.push(...breakNodes);
@@ -760,7 +789,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         [ast.consequent.loc?.start.line],
         []
       );
-      this.connectParents([node], [emptyChildNode])
+      this.connectParents([node], [emptyChildNode]);
       trueNodes.push(emptyChildNode);
 
       this.cfg.edges[count].branchType = true;
@@ -777,9 +806,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     // Visit false flow
     if (ast.alternate) {
       count = this.cfg.edges.length;
-      const { childNodes, breakNodes } = this.visitChild(ast.alternate, [
-        node,
-      ]);
+      const { childNodes, breakNodes } = this.visitChild(ast.alternate, [node]);
       const falseNodes = childNodes;
       totalBreakNodes.push(...breakNodes);
 
@@ -826,6 +853,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     }
   }
 
+<<<<<<< HEAD
 
   private visitConditional(ast: any, parents: Node[]): ReturnValue {
     const node: BranchNode = this.createBranchNode(
@@ -835,6 +863,17 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         type: ast.test?.type,
         operator: ast.test?.operator,
       });
+=======
+  private visitConditional(ast: t.Conditional, parents: Node[]): ReturnValue {
+    const node: BranchNode = this.createBranchNode([ast.loc.start.line], [], {
+      type: ast.test.type,
+      operator:
+        ast.test.type === "BinaryExpression" ||
+        ast.test.type === "UnaryExpression"
+          ? ast.test.operator
+          : "",
+    });
+>>>>>>> develop
     this.connectParents(parents, [node]);
 
     // Store all break points
@@ -842,9 +881,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
 
     // Visit true flow
     let count = this.cfg.edges.length;
-    const { childNodes, breakNodes } = this.visitChild(ast.consequent,
-      [node]
-    );
+    const { childNodes, breakNodes } = this.visitChild(ast.consequent, [node]);
     const trueNodes = childNodes;
     totalBreakNodes.push(...breakNodes);
 
@@ -870,10 +907,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     // Visit false flow
     if (ast.alternate) {
       count = this.cfg.edges.length;
-      const { childNodes, breakNodes } = this.visitChild(
-        ast.alternate,
-        [node]
-      );
+      const { childNodes, breakNodes } = this.visitChild(ast.alternate, [node]);
       const falseNodes = childNodes;
       totalBreakNodes.push(...breakNodes);
 
@@ -917,8 +951,9 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     }
   }
 
-  private visitTryStatement(ast: any, parents:Node[]): ReturnValue {
+  private visitTryStatement(ast: t.TryStatement, parents: Node[]): ReturnValue {
     // TODO finalizer try -> catch -> finally
+<<<<<<< HEAD
     const node: BranchNode = this.createBranchNode(
       [ast.loc?.start.line],
       [],
@@ -927,6 +962,12 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         operator: 'exception',
       }
     );
+=======
+    const node: BranchNode = this.createBranchNode([ast.loc.start.line], [], {
+      type: ast.type,
+      operator: "exception",
+    });
+>>>>>>> develop
 
     this.connectParents(parents, [node]);
 
@@ -935,9 +976,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
 
     // Visit try flow
     let count = this.cfg.edges.length;
-    let tryNodes = this.visitChild(ast.block, [
-      node,
-    ]);
+    const tryNodes = this.visitChild(ast.block, [node]);
     const tryChildNodes = tryNodes.childNodes;
     totalBreakNodes.push(...tryNodes.breakNodes);
 
@@ -948,7 +987,11 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     } else {
       // Add empty placeholder node
       const emptyChildNode = this.createPlaceholderNode(
+<<<<<<< HEAD
         [ast.consequent.loc?.start.line],
+=======
+        [ast.block.loc.start.line],
+>>>>>>> develop
         []
       );
       tryChildNodes.push(emptyChildNode);
@@ -962,9 +1005,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
 
     // Visit catch flow
     count = this.cfg.edges.length;
-    const catchNodes = this.visitChild(ast.handler, [
-      node,
-    ]);
+    const catchNodes = this.visitChild(ast.handler, [node]);
     const catchChildNodes = catchNodes.childNodes;
     totalBreakNodes.push(...catchNodes.breakNodes);
 
@@ -974,6 +1015,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
       this.cfg.edges[count].branchType = false;
     } else {
       // Add empty placeholder node
+<<<<<<< HEAD
       let emptyChildNode;
       if (!ast || !ast.alternate || !ast.loc) {
         emptyChildNode = this.createPlaceholderNode(
@@ -986,6 +1028,12 @@ export class ControlFlowGraphGenerator implements CFGFactory {
             []
         );
       }
+=======
+      const emptyChildNode = this.createPlaceholderNode(
+        [ast.handler.loc.start.line],
+        []
+      );
+>>>>>>> develop
       catchChildNodes.push(emptyChildNode);
 
       this.cfg.edges.push({
@@ -1001,8 +1049,13 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
+<<<<<<< HEAD
   private visitCatchClause(ast: any, parents:Node[]): ReturnValue {
     const node: Node = this.createNode([ast.loc?.start.line], []);
+=======
+  private visitCatchClause(ast: t.CatchClause, parents: Node[]): ReturnValue {
+    const node: Node = this.createNode([ast.loc.start.line], []);
+>>>>>>> develop
     this.connectParents(parents, [node]);
 
     return {
@@ -1011,10 +1064,24 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
+<<<<<<< HEAD
   private visitWhileStatement(ast: any, parents: Node[]): ReturnValue {
     const node: Node = this.createBranchNode([ast.loc?.start.line], [], {
       type: ast.test?.type,
       operator: ast.test.operator,
+=======
+  private visitWhileStatement(
+    ast: t.WhileStatement,
+    parents: Node[]
+  ): ReturnValue {
+    const node: Node = this.createBranchNode([ast.loc.start.line], [], {
+      type: ast.test.type,
+      operator:
+        ast.test.type === "BinaryExpression" ||
+        ast.test.type === "UnaryExpression"
+          ? ast.test.operator
+          : "",
+>>>>>>> develop
     });
     this.connectParents(parents, [node]);
 
@@ -1042,7 +1109,11 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     }
 
     // Add empty placeholder node for the false flow
+<<<<<<< HEAD
     const falseNode = this.createPlaceholderNode( [ast.loc?.end.line], []);
+=======
+    const falseNode = this.createPlaceholderNode([ast.loc.end.line], []);
+>>>>>>> develop
     this.cfg.edges.push({
       from: node.id,
       to: falseNode.id,
@@ -1066,10 +1137,20 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
+<<<<<<< HEAD
   private visitForOfStatement(ast: any, parents: Node[]): ReturnValue {
     const node: Node = this.createBranchNode([ast.loc?.start.line], [], {
       type: 'CallExpression',
       operator: 'isEmpty',
+=======
+  private visitForOfStatement(
+    ast: t.ForOfStatement | t.ForInStatement,
+    parents: Node[]
+  ): ReturnValue {
+    const node: Node = this.createBranchNode([ast.loc.start.line], [], {
+      type: "CallExpression",
+      operator: "isEmpty",
+>>>>>>> develop
     });
     this.connectParents(parents, [node]);
 
@@ -1128,8 +1209,16 @@ export class ControlFlowGraphGenerator implements CFGFactory {
    * @constructor
    * @private
    */
+<<<<<<< HEAD
   private visitBreakStatement(ast: any, parents: Node[]): ReturnValue {
     const node: Node = this.createNode([ast.loc?.start.line], []);
+=======
+  private visitBreakStatement(
+    ast: t.BreakStatement,
+    parents: Node[]
+  ): ReturnValue {
+    const node: Node = this.createNode([ast.loc.start.line], []);
+>>>>>>> develop
     this.connectParents(parents, [node]);
 
     return {
@@ -1139,8 +1228,16 @@ export class ControlFlowGraphGenerator implements CFGFactory {
   }
 
   // TODO currently incorrect is passthrough now
+<<<<<<< HEAD
   private visitContinueStatement(ast: any, parents: Node[]): ReturnValue {
     const node: Node = this.createNode([ast.loc?.start.line], []);
+=======
+  private visitContinueStatement(
+    ast: t.ContinueStatement,
+    parents: Node[]
+  ): ReturnValue {
+    const node: Node = this.createNode([ast.loc.start.line], []);
+>>>>>>> develop
     this.connectParents(parents, [node]);
 
     return {
@@ -1149,8 +1246,12 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
-  private visitDoWhileStatement(ast: any, parents: Node[]): ReturnValue {
+  private visitDoWhileStatement(
+    ast: t.DoWhileStatement,
+    parents: Node[]
+  ): ReturnValue {
     // entry node
+<<<<<<< HEAD
     const entryNode: Node = this.createBranchNode(
       [ast.loc?.start.line],
       [],
@@ -1159,16 +1260,25 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         operator: ast.test.operator,
       }
     );
+=======
+    const entryNode: Node = this.createBranchNode([ast.loc.start.line], [], {
+      type: ast.test.type,
+      operator:
+        ast.test.type === "BinaryExpression" ||
+        ast.test.type === "UnaryExpression"
+          ? ast.test.operator
+          : "",
+    });
+>>>>>>> develop
     this.connectParents(parents, [entryNode]);
 
     // TODO: We can check if a node is generated. This eliminates the need for entryNode
     // 'do' block
-    const { childNodes, breakNodes } = this.visitChild(ast.body, [
-      entryNode,
-    ]);
+    const { childNodes, breakNodes } = this.visitChild(ast.body, [entryNode]);
     const trueNodes = childNodes;
 
     // while check
+<<<<<<< HEAD
     const whileNode: Node = this.createBranchNode(
       [ast.loc?.start.line],
       [],
@@ -1177,6 +1287,16 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         operator: ast.test.operator,
       }
     );
+=======
+    const whileNode: Node = this.createBranchNode([ast.loc.start.line], [], {
+      type: ast.test.type,
+      operator:
+        ast.test.type === "BinaryExpression" ||
+        ast.test.type === "UnaryExpression"
+          ? ast.test.operator
+          : "",
+    });
+>>>>>>> develop
     this.connectParents(trueNodes, [whileNode]);
 
     // Connect back to the entry node and mark as true branch
@@ -1187,10 +1307,14 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     });
 
     // Add empty placeholder node for the false flow
+<<<<<<< HEAD
     const falseNode: Node = this.createPlaceholderNode(
       [ast.loc?.end.line],
       []
     );
+=======
+    const falseNode: Node = this.createPlaceholderNode([ast.loc.end.line], []);
+>>>>>>> develop
     this.cfg.edges.push({
       from: whileNode.id,
       to: falseNode.id,
@@ -1211,21 +1335,30 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
-  private visitSwitchStatement(ast: any, parents: Node[]): ReturnValue {
+  private visitSwitchStatement(
+    ast: t.SwitchStatement,
+    parents: Node[]
+  ): ReturnValue {
     // TODO currently incorrect
+<<<<<<< HEAD
     const node: Node = this.createBranchNode([ast.loc?.start.line], [], {
       type: 'Switch',
       operator: '==',
+=======
+    const node: Node = this.createBranchNode([ast.loc.start.line], [], {
+      type: "Switch",
+      operator: "==",
+>>>>>>> develop
     });
     this.connectParents(parents, [node]);
 
     // Add empty placeholder node for the false flow
     const falseNode = this.createPlaceholderNode([ast.loc?.end.line], []);
 
-    let nodes = [node]
+    let nodes = [node];
     for (const switchCase of ast.cases) {
       const { childNodes, breakNodes } = this.visitChild(switchCase, nodes);
-      nodes = childNodes
+      nodes = childNodes;
 
       // Connect break points
       for (const breakNode of breakNodes) {
@@ -1249,25 +1382,32 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
+<<<<<<< HEAD
   private visitSwitchCase(ast: any, parents: Node[]): ReturnValue {
     const node: Node = this.createBranchNode([ast.loc?.start.line], [], {
       type: 'switchCase',
       operator: '==',
+=======
+  private visitSwitchCase(ast: t.SwitchCase, parents: Node[]): ReturnValue {
+    const node: Node = this.createBranchNode([ast.loc.start.line], [], {
+      type: "switchCase",
+      operator: "==",
+>>>>>>> develop
     });
     this.connectParents(parents, [node]);
 
     const count = this.cfg.edges.length;
 
-    let nodes = [node]
-    const totalBreakNodes = []
+    let nodes = [node];
+    const totalBreakNodes = [];
     for (const child of ast.consequent) {
       const { childNodes, breakNodes } = this.visitChild(child, nodes);
-      nodes = childNodes
-      totalBreakNodes.push(...breakNodes)
+      nodes = childNodes;
+      totalBreakNodes.push(...breakNodes);
     }
 
     const trueNodes = nodes;
-    const childNodes = []
+    const childNodes = [];
 
     // Check if a child node was created
     if (this.cfg.edges[count]) {
@@ -1287,7 +1427,7 @@ export class ControlFlowGraphGenerator implements CFGFactory {
         branchType: true,
       });
 
-      childNodes.push(emptyChildNode)
+      childNodes.push(emptyChildNode);
     }
 
     // Add empty placeholder node for the false flow
@@ -1304,10 +1444,21 @@ export class ControlFlowGraphGenerator implements CFGFactory {
     };
   }
 
+<<<<<<< HEAD
   private visitForStatement(ast: any, parents: Node[]): ReturnValue {
     const node: Node = this.createBranchNode([ast.loc?.start.line], [], {
       type: ast.test?.type,
       operator: ast.test.operator || ast.test.name || ast.test.value,
+=======
+  private visitForStatement(ast: t.ForStatement, parents: Node[]): ReturnValue {
+    const node: Node = this.createBranchNode([ast.loc.start.line], [], {
+      type: ast.test.type,
+      operator:
+        ast.test.type === "BinaryExpression" ||
+        ast.test.type === "UnaryExpression"
+          ? ast.test.operator
+          : "",
+>>>>>>> develop
     });
     this.connectParents(parents, [node]);
     // TODO For each probably not supported
