@@ -19,6 +19,7 @@
 import { traverse } from "@babel/core";
 import { Export, ExportVisitor } from "./ExportVisitor";
 import { IdentifierVisitor } from "./IdentifierVisitor";
+import * as t from "@babel/types";
 
 /**
  * Exports generator for targets.
@@ -32,11 +33,14 @@ export class ExportGenerator {
    * @param targetPath The path of the AST
    * @param targetAST The AST of the target
    */
-  generate(targetPath: string, targetAST: any): Export[] {
-    const identifierVisitor = new IdentifierVisitor(targetPath)
-    traverse(targetAST, identifierVisitor)
+  generate(targetPath: string, targetAST: t.Node): Export[] {
+    const identifierVisitor = new IdentifierVisitor(targetPath);
+    traverse(targetAST, identifierVisitor);
 
-    const visitor = new ExportVisitor(targetPath, identifierVisitor.identifiers);
+    const visitor = new ExportVisitor(
+      targetPath,
+      identifierVisitor.identifiers
+    );
 
     traverse(targetAST, visitor);
 
