@@ -21,62 +21,64 @@ export enum ExportType {
   function,
   class,
   const,
-  unknown
+  unknown,
 }
 
 export class IdentifierVisitor extends Visitor {
-  private _identifiers: Map<string, ExportType>
+  private _identifiers: Map<string, ExportType>;
 
   get identifiers(): Map<string, ExportType> {
     return this._identifiers;
   }
 
   constructor(filePath: string) {
-    super(filePath)
-    this._identifiers = new Map<string, ExportType>()
+    super(filePath);
+    this._identifiers = new Map<string, ExportType>();
   }
 
   // identifiable stuff
   public FunctionDeclaration: (path) => void = (path) => {
     let identifier = path.node.id?.name;
 
-    if (path.parent.type === 'ObjectPattern') {
-      identifier = path.parent.key.name
-    } else if (path.parent.type === 'ExportDefaultDeclaration') {
-      identifier = 'default'
+    if (path.parent.type === "ObjectPattern") {
+      identifier = path.parent.key.name;
+    } else if (path.parent.type === "ExportDefaultDeclaration") {
+      identifier = "default";
     }
 
     if (!identifier) {
-      throw new Error(`Unsupported identifier declaration: ${path.parent.type}`)
+      throw new Error(
+        `Unsupported identifier declaration: ${path.parent.type}`
+      );
     }
 
-    this._identifiers.set(identifier, ExportType.function)
-  }
+    this._identifiers.set(identifier, ExportType.function);
+  };
 
   public ClassDeclaration: (path) => void = (path) => {
     const identifier = path.node.id.name;
-    this._identifiers.set(identifier, ExportType.class)
-  }
+    this._identifiers.set(identifier, ExportType.class);
+  };
 
   public VariableDeclaration: (path) => void = (path) => {
     for (const declaration of path.node.declarations) {
       const identifier = declaration.id.name;
 
-      if (declaration.id.type === 'ObjectPattern') {
+      if (declaration.id.type === "ObjectPattern") {
         // TODO support this
-        continue
+        continue;
       }
 
       if (!declaration.init) {
-        this._identifiers.set(identifier, ExportType.unknown)
-      } else if (declaration.init.type === "ArrowFunctionExpression"
-        || declaration.init.type === "FunctionExpression") {
-        this._identifiers.set(identifier, ExportType.function) // not always the case
+        this._identifiers.set(identifier, ExportType.unknown);
+      } else if (
+        declaration.init.type === "ArrowFunctionExpression" ||
+        declaration.init.type === "FunctionExpression"
+      ) {
+        this._identifiers.set(identifier, ExportType.function); // not always the case
       } else {
-        this._identifiers.set(identifier, ExportType.const) // not always the case
+        this._identifiers.set(identifier, ExportType.const); // not always the case
       }
     }
-  }
+  };
 }
-
-
