@@ -21,19 +21,18 @@ export class ImportVisitor extends Visitor {
   private _imports: Set<string>;
 
   constructor(filePath) {
-    super(filePath)
+    super(filePath);
     this._imports = new Set<string>();
   }
 
   public ImportDeclaration: (path) => void = (path) => {
-    this._imports.add(path.node.source.value)
+    this._imports.add(path.node.source.value);
   };
 
   public CallExpression: (path) => void = (path) => {
-    if (path.node.callee.name === 'require') {
-
-      if (path.node.arguments[0].type === 'StringLiteral') {
-        this._imports.add(path.node.arguments[0].value)
+    if (path.node.callee.name === "require") {
+      if (path.node.arguments[0].type === "StringLiteral") {
+        this._imports.add(path.node.arguments[0].value);
       } else {
         // This tool does not support dynamic require statements.
         // throw new Error("This tool does not support dynamic require statements.")
