@@ -21,248 +21,249 @@ import { Element } from "./Element";
 // TODO add missing
 
 export interface Relation {
-  relation: RelationType
-  involved: Element[]
-  computed?: boolean
+  relation: RelationType;
+  involved: Element[];
+  computed?: boolean;
 }
 
 export enum RelationType {
   // Primary Expressions
   // TODO others
-  FunctionDefinition="L_R",
-  ClassDefinition="class L",
+  FunctionDefinition = "L_R",
+  ClassDefinition = "class L",
 
-
-  Await="await L",
+  Await = "await L",
 
   // object
-  Object="{L}", // TODO not correct L doesnt matter
-  ObjectProperty="L:R",
+  Object = "{L}", // TODO not correct L doesnt matter
+  ObjectProperty = "L:R",
 
   // array
-  Array="[L]",
-  Sequence="(L,R)",
-
-
-
-
+  Array = "[L]",
+  Sequence = "(L,R)",
 
   // Left-hand-side Expressions
-  PropertyAccessor="L.R",
-  New="new L()",
+  PropertyAccessor = "L.R",
+  New = "new L()",
   // TODO new.target
   // TODO import.meta
   // TODO super
-  Spread="...L",
+  Spread = "...L",
 
   // UNARY
   // Increment and Decrement
-  PlusPlusPostFix="L++",
-  MinusMinusPostFix="L--",
-  PlusPlusPrefix="++L",
-  MinusMinusPrefix="--L",
+  PlusPlusPostFix = "L++",
+  MinusMinusPostFix = "L--",
+  PlusPlusPrefix = "++L",
+  MinusMinusPrefix = "--L",
 
   // Unary
-  Delete="delete L",
-  Void="void L",
-  TypeOf="typeof L",
-  PlusUnary="+L",
-  MinusUnary="-L",
-  BitwiseNotUnary="~L",
-  LogicalNotUnary="!L",
+  Delete = "delete L",
+  Void = "void L",
+  TypeOf = "typeof L",
+  PlusUnary = "+L",
+  MinusUnary = "-L",
+  BitwiseNotUnary = "~L",
+  LogicalNotUnary = "!L",
 
   // BINARY
   // Arithmetic
-  Addition="L+R",
-  Subtraction="L-R",
-  Division="L/R",
-  Multiplication="L*R",
-  Remainder="L%R",
-  Exponentiation="L**R",
+  Addition = "L+R",
+  Subtraction = "L-R",
+  Division = "L/R",
+  Multiplication = "L*R",
+  Remainder = "L%R",
+  Exponentiation = "L**R",
 
   // Relation
-  In="L in R",
-  InstanceOf="L instanceof R",
-  Less="L<R",
-  Greater="L>R",
-  LessOrEqual="L<=R",
-  GreaterOrEqual="L>=R",
+  In = "L in R",
+  InstanceOf = "L instanceof R",
+  Less = "L<R",
+  Greater = "L>R",
+  LessOrEqual = "L<=R",
+  GreaterOrEqual = "L>=R",
 
   // Equality
-  Equality="L==R",
-  InEquality="L!=R",
-  StrictEquality="L===R",
-  StrictInequality="L!==R",
+  Equality = "L==R",
+  InEquality = "L!=R",
+  StrictEquality = "L===R",
+  StrictInequality = "L!==R",
 
   // Bitwise shift
-  BitwiseLeftShift="L<<R",
-  BitwiseRightShift="L>>R",
-  BitwiseUnsignedRightShift="L>>>R",
+  BitwiseLeftShift = "L<<R",
+  BitwiseRightShift = "L>>R",
+  BitwiseUnsignedRightShift = "L>>>R",
 
   // Binary bitwise
-  BitwiseAnd="L&R",
-  BitwiseOr="L|R",
-  BitwiseXor="L^R",
+  BitwiseAnd = "L&R",
+  BitwiseOr = "L|R",
+  BitwiseXor = "L^R",
 
   // Binary logical
-  LogicalAnd="L&&R",
-  LogicalOr="L||R",
-  NullishCoalescing="L??R",
+  LogicalAnd = "L&&R",
+  LogicalOr = "L||R",
+  NullishCoalescing = "L??R",
 
   // Ternary
-  Conditional="C?L:R",
+  Conditional = "C?L:R",
 
   // Optional chaining
   // TODO
 
   // Assignment
-  Assignment="L=R",
-  MultiplicationAssignment="L*=R",
-  ExponentiationAssignment="L**=R",
-  DivisionAssignment="L/=R",
-  RemainderAssigment="L%=R",
-  AdditionAssignment="L+=R",
-  SubtractionAssignment="L-=R",
-  LeftShiftAssignment="L<<=R",
-  RightShiftAssignment="L>>=R",
-  UnSignedRightShiftAssignment="L>>>=R",
-  BitwiseAndAssignment="L&=R",
-  BitwiseXorAssignment="L^=R",
-  BitwiseOrAssignment="L|=R",
-  LogicalAndAssignment="L&&=R",
-  LogicalOrAssignment="L||=R",
-  LogicalNullishAssignment="L??=R",
+  Assignment = "L=R",
+  MultiplicationAssignment = "L*=R",
+  ExponentiationAssignment = "L**=R",
+  DivisionAssignment = "L/=R",
+  RemainderAssigment = "L%=R",
+  AdditionAssignment = "L+=R",
+  SubtractionAssignment = "L-=R",
+  LeftShiftAssignment = "L<<=R",
+  RightShiftAssignment = "L>>=R",
+  UnSignedRightShiftAssignment = "L>>>=R",
+  BitwiseAndAssignment = "L&=R",
+  BitwiseXorAssignment = "L^=R",
+  BitwiseOrAssignment = "L|=R",
+  LogicalAndAssignment = "L&&=R",
+  LogicalOrAssignment = "L||=R",
+  LogicalNullishAssignment = "L??=R",
   // TODO destructuring assignment
 
-
   // function
-  Return="L->R",
+  Return = "L->R",
 
   // MULTI
   // function
-  Call="L(R)",
+  Call = "L(R)",
 
-  PrivateName="#L"
+  PrivateName = "#L",
 }
 
-
-export function getRelationType(type: string, operator: string, prefix=false): RelationType {
+export function getRelationType(
+  type: string,
+  operator: string,
+  prefix = false
+): RelationType {
   if (type === "unary") {
     switch (operator) {
       case "++":
-        return prefix ? RelationType.PlusPlusPrefix : RelationType.PlusPlusPostFix
+        return prefix
+          ? RelationType.PlusPlusPrefix
+          : RelationType.PlusPlusPostFix;
       case "--":
-        return prefix ? RelationType.MinusMinusPrefix : RelationType.MinusMinusPostFix
+        return prefix
+          ? RelationType.MinusMinusPrefix
+          : RelationType.MinusMinusPostFix;
 
       case "delete":
-        return RelationType.Delete
+        return RelationType.Delete;
       case "void":
-        return RelationType.Void
+        return RelationType.Void;
       case "typeof":
-        return RelationType.TypeOf
+        return RelationType.TypeOf;
       case "+":
-        return RelationType.PlusUnary
+        return RelationType.PlusUnary;
       case "-":
-        return RelationType.MinusUnary
+        return RelationType.MinusUnary;
       case "~":
-        return RelationType.BitwiseNotUnary
+        return RelationType.BitwiseNotUnary;
       case "!":
-        return RelationType.LogicalNotUnary
+        return RelationType.LogicalNotUnary;
     }
   } else if (type === "binary") {
     switch (operator) {
       case "+":
-        return RelationType.Addition
+        return RelationType.Addition;
       case "-":
-        return RelationType.Subtraction
+        return RelationType.Subtraction;
       case "/":
-        return RelationType.Division
+        return RelationType.Division;
       case "*":
-        return RelationType.Multiplication
+        return RelationType.Multiplication;
       case "%":
-        return RelationType.Remainder
+        return RelationType.Remainder;
       case "**":
-        return RelationType.Exponentiation
+        return RelationType.Exponentiation;
 
       case "in":
-        return RelationType.In
+        return RelationType.In;
       case "instanceof":
-        return RelationType.InstanceOf
+        return RelationType.InstanceOf;
       case "<":
-        return RelationType.Less
+        return RelationType.Less;
       case ">":
-        return RelationType.Greater
+        return RelationType.Greater;
       case "<=":
-        return RelationType.LessOrEqual
+        return RelationType.LessOrEqual;
       case ">=":
-        return RelationType.GreaterOrEqual
+        return RelationType.GreaterOrEqual;
 
       case "==":
-        return RelationType.Equality
+        return RelationType.Equality;
       case "!=":
-        return RelationType.InEquality
+        return RelationType.InEquality;
       case "===":
-        return RelationType.StrictEquality
+        return RelationType.StrictEquality;
       case "!==":
-        return RelationType.StrictInequality
+        return RelationType.StrictInequality;
 
       case "<<":
-        return RelationType.BitwiseLeftShift
+        return RelationType.BitwiseLeftShift;
       case ">>":
-        return RelationType.BitwiseRightShift
+        return RelationType.BitwiseRightShift;
       case ">>>":
-        return RelationType.BitwiseUnsignedRightShift
+        return RelationType.BitwiseUnsignedRightShift;
 
       case "&":
-        return RelationType.BitwiseAnd
+        return RelationType.BitwiseAnd;
       case "|":
-        return RelationType.BitwiseOr
+        return RelationType.BitwiseOr;
       case "^":
-        return RelationType.BitwiseXor
+        return RelationType.BitwiseXor;
 
       case "&&":
-        return RelationType.LogicalAnd
+        return RelationType.LogicalAnd;
       case "||":
-        return RelationType.LogicalAnd
+        return RelationType.LogicalAnd;
       case "??":
-        return RelationType.NullishCoalescing
+        return RelationType.NullishCoalescing;
     }
   } else if (type === "assignment") {
     switch (operator) {
       case "=":
-        return RelationType.Assignment
+        return RelationType.Assignment;
       case "*=":
-        return RelationType.MultiplicationAssignment
+        return RelationType.MultiplicationAssignment;
       case "**=":
-        return RelationType.ExponentiationAssignment
+        return RelationType.ExponentiationAssignment;
       case "/=":
-        return RelationType.DivisionAssignment
+        return RelationType.DivisionAssignment;
       case "%=":
-        return RelationType.RemainderAssigment
+        return RelationType.RemainderAssigment;
       case "+=":
-        return RelationType.AdditionAssignment
+        return RelationType.AdditionAssignment;
       case "-=":
-        return RelationType.SubtractionAssignment
+        return RelationType.SubtractionAssignment;
       case "<<=":
-        return RelationType.LeftShiftAssignment
+        return RelationType.LeftShiftAssignment;
       case ">>=":
-        return RelationType.RightShiftAssignment
+        return RelationType.RightShiftAssignment;
       case ">>>=":
-        return RelationType.UnSignedRightShiftAssignment
+        return RelationType.UnSignedRightShiftAssignment;
       case "&=":
-        return RelationType.BitwiseAndAssignment
+        return RelationType.BitwiseAndAssignment;
       case "^=":
-        return RelationType.BitwiseXorAssignment
+        return RelationType.BitwiseXorAssignment;
       case "|=":
-        return RelationType.BitwiseOrAssignment
+        return RelationType.BitwiseOrAssignment;
       case "&&=":
-        return RelationType.LogicalAndAssignment
+        return RelationType.LogicalAndAssignment;
       case "||=":
-        return RelationType.LogicalOrAssignment
+        return RelationType.LogicalOrAssignment;
       case "??=":
-        return RelationType.LogicalNullishAssignment
+        return RelationType.LogicalNullishAssignment;
     }
   }
 
-  throw new Error(`Unsupported relation type operator: ${type} -> ${operator}`)
+  throw new Error(`Unsupported relation type operator: ${type} -> ${operator}`);
 }
