@@ -20,6 +20,7 @@ import { traverse } from "@babel/core";
 import { ObjectVisitor } from "./ObjectVisitor";
 import { ComplexObject } from "./ComplexObject";
 import { Export } from "../../../dependency/ExportVisitor";
+import * as t from "@babel/types";
 
 /**
  * Typing generator for targets.
@@ -33,7 +34,11 @@ export class ObjectGenerator {
    * @param filePath the path of the current file
    * @param targetAST The AST of the target
    */
-  generate(filePath: string, targetAST: any, exports: Export[]): ComplexObject[] {
+  generate(
+    filePath: string,
+    targetAST: t.Node,
+    exports: Export[]
+  ): ComplexObject[] {
     const visitor = new ObjectVisitor(filePath, exports);
 
     traverse(targetAST, visitor);
