@@ -27,28 +27,28 @@ import { Scope } from "../discovery/Scope";
  * @author Dimitri Stallenberg
  */
 export abstract class TypeResolver {
-  private _relationTyping: Map<Relation, TypeProbability>
-  private _elementTyping: Map<Element, TypeProbability>
+  private _relationTyping: Map<Relation, TypeProbability>;
+  private _elementTyping: Map<Element, TypeProbability>;
 
-  private _relationFullyResolved: Set<Relation>
+  private _relationFullyResolved: Set<Relation>;
 
-  private _wrapperElementIsRelation: Map<string, Relation>
+  private _wrapperElementIsRelation: Map<string, Relation>;
 
-  private processed: Map<Relation, Set<Element>>
+  private processed: Map<Relation, Set<Element>>;
 
   /**
    * Constructor
    */
   constructor() {
-    this._relationTyping = new Map()
-    this._elementTyping = new Map()
-    this._relationFullyResolved = new Set()
+    this._relationTyping = new Map();
+    this._elementTyping = new Map();
+    this._relationFullyResolved = new Set();
 
-    this.processed = new Map<Relation, Set<Element>>()
+    this.processed = new Map<Relation, Set<Element>>();
   }
 
   get availableTypes() {
-    return [...this.elementTyping.values()]//, ...this.relationTyping.values()])
+    return [...this.elementTyping.values()]; //, ...this.relationTyping.values()])
   }
 
   /**
@@ -59,14 +59,19 @@ export abstract class TypeResolver {
    * @param wrapperElementIsRelation a map that specifies which elements are which relations
    * @param objects the user defined objects that are available to the file under evaluation
    */
-  abstract resolveTypes(elements: Element[], relations: Relation[], wrapperElementIsRelation: Map<string, Relation>, objects: ComplexObject[])
+  abstract resolveTypes(
+    elements: Element[],
+    relations: Relation[],
+    wrapperElementIsRelation: Map<string, Relation>,
+    objects: ComplexObject[]
+  );
 
   /**
    * Returns the identifierDescription of the variable in the given scope
    * @param scope the scope the variable is in
    * @param variableName the name of the variable
    */
-  abstract getTyping(scope: Scope, variableName: string): TypeProbability
+  abstract getTyping(scope: Scope, variableName: string): TypeProbability;
 
   /**
    * Sets the identifierDescription of the specified relation
@@ -74,109 +79,142 @@ export abstract class TypeResolver {
    * @param type the identifierDescription of the relation
    * @param value the score of identifierDescription (higher score means higher probability)
    */
-  setRelationType(relation: Relation, type: string | TypeProbability, value: number) {
+  setRelationType(
+    relation: Relation,
+    type: string | TypeProbability,
+    value: number
+  ) {
     if (relation.relation === RelationType.PropertyAccessor) {
-      this.setElementType(relation, relation.involved[1], type, 1)
-      this.setProcessed(relation, relation.involved[1])
+      this.setElementType(relation, relation.involved[1], type, 1);
+      this.setProcessed(relation, relation.involved[1]);
     }
 
     if (this.relationTyping.has(relation)) {
-      const probabilities = this.relationTyping.get(relation)
-      probabilities.addType(type, value)
+      const probabilities = this.relationTyping.get(relation);
+      probabilities.addType(type, value);
     } else {
-      const probabilities = new TypeProbability()
-      this.relationTyping.set(relation, probabilities)
-      probabilities.addType(type, value)
+      const probabilities = new TypeProbability();
+      this.relationTyping.set(relation, probabilities);
+      probabilities.addType(type, value);
     }
   }
 
-  setElementTypeToElement(relation: Relation, element: Element, typeElement: Element, value: number) {
+  setElementTypeToElement(
+    relation: Relation,
+    element: Element,
+    typeElement: Element,
+    value: number
+  ) {
     if (!this.elementTyping.has(typeElement)) {
-      this.elementTyping.set(typeElement, new TypeProbability())
+      this.elementTyping.set(typeElement, new TypeProbability());
     }
 
     if (typeElement.type === ElementType.Relation) {
-      const relation = this._wrapperElementIsRelation.get(typeElement.value)
+      const relation = this._wrapperElementIsRelation.get(typeElement.value);
 
       if (!relation) {
-        throw new Error(`Cannot find relation: ${typeElement.value}`)
+        throw new Error(`Cannot find relation: ${typeElement.value}`);
       }
 
-      this.setElementType(relation, element, this.relationTyping.get(relation), value)
-
+      this.setElementType(
+        relation,
+        element,
+        this.relationTyping.get(relation),
+        value
+      );
     } else {
-      this.setElementType(relation, element, this.elementTyping.get(typeElement), value)
+      this.setElementType(
+        relation,
+        element,
+        this.elementTyping.get(typeElement),
+        value
+      );
     }
   }
 
-  setElementType(relation: Relation, element: Element, type: string | TypeProbability, value: number, object: ComplexObject = null, propertyTypings: Map<string, TypeProbability> = null) {
+  setElementType(
+    relation: Relation,
+    element: Element,
+    type: string | TypeProbability,
+    value: number,
+    object: ComplexObject = null,
+    propertyTypings: Map<string, TypeProbability> = null
+  ) {
     if (!this.processed.has(relation)) {
-      this.processed.set(relation, new Set<Element>())
+      this.processed.set(relation, new Set<Element>());
     }
 
     if (this.processed.get(relation).has(element)) {
-      return
+      return;
     }
 
-    this._setElementType(element, type, value, object, propertyTypings)
+    this._setElementType(element, type, value, object, propertyTypings);
   }
 
   setProcessed(relation: Relation, element: Element) {
     if (!this.processed.has(relation)) {
-      this.processed.set(relation, new Set<Element>())
+      this.processed.set(relation, new Set<Element>());
     }
 
-    this.processed.get(relation).add(element)
+    this.processed.get(relation).add(element);
   }
 
-    /**
+  /**
    * Sets the identifierDescription of the specified element
    * @param element the element to set the identifierDescription of
    * @param type the type of the element
    * @param value the score of type (higher score means higher probability)
    */
-  _setElementType(element: Element, type: string | TypeProbability, value: number, object: ComplexObject = null, propertyTypings: Map<string, TypeProbability> = null) {
+  _setElementType(
+    element: Element,
+    type: string | TypeProbability,
+    value: number,
+    object: ComplexObject = null,
+    propertyTypings: Map<string, TypeProbability> = null
+  ) {
     if (element.type === ElementType.Relation) {
-      const relation = this._wrapperElementIsRelation.get(element.value)
+      const relation = this._wrapperElementIsRelation.get(element.value);
 
       if (!relation) {
-        throw new Error(`Cannot find relation: ${element.value}`)
+        throw new Error(`Cannot find relation: ${element.value}`);
       }
 
-      this.setRelationType(relation, type, 1)
+      this.setRelationType(relation, type, 1);
     }
 
     if (!this.elementTyping.has(element)) {
-      this.elementTyping.set(element, new TypeProbability())
+      this.elementTyping.set(element, new TypeProbability());
     }
 
-    const typeMap = this.elementTyping.get(element)
-    typeMap.addType(type, value, object, propertyTypings)
+    const typeMap = this.elementTyping.get(element);
+    typeMap.addType(type, value, object, propertyTypings);
   }
 
   getRelationType(relation: Relation) {
     if (!this.relationTyping.has(relation)) {
-      this.relationTyping.set(relation, new TypeProbability())
+      this.relationTyping.set(relation, new TypeProbability());
     }
 
-    return this.relationTyping.get(relation)
+    return this.relationTyping.get(relation);
   }
 
   getElementType(element: Element) {
-    if (element.type === 'relation') {
-      const relation: Relation = this.wrapperElementIsRelation.get(element.value)
-      return this.getRelationType(relation)
+    if (element.type === "relation") {
+      const relation: Relation = this.wrapperElementIsRelation.get(
+        element.value
+      );
+      return this.getRelationType(relation);
     }
 
     if (!this.elementTyping.has(element)) {
-      this.elementTyping.set(element, new TypeProbability())
+      this.elementTyping.set(element, new TypeProbability());
     }
 
-    return this.elementTyping.get(element)
+    return this.elementTyping.get(element);
   }
 
   get elements() {
-    return [...this.elementTyping.keys()]
+    return [...this.elementTyping.keys()];
   }
 
   private get relationTyping(): Map<Relation, TypeProbability> {
@@ -190,7 +228,6 @@ export abstract class TypeResolver {
   get relationFullyResolved(): Set<Relation> {
     return this._relationFullyResolved;
   }
-
 
   get wrapperElementIsRelation(): Map<string, Relation> {
     return this._wrapperElementIsRelation;
