@@ -15,71 +15,85 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import * as t from "@babel/types";
 
 export class BranchDistance {
-  private K = 1 // punishment factor
+  private K = 1; // punishment factor
 
   public static branchDistance(
     condition: string,
     condition_ast: string,
-    variables: any,
+    variables: Record<string, unknown>,
     target: boolean
   ): number {
-    if (condition === undefined || condition_ast === undefined || variables === undefined) {
-      return 1
+    if (
+      condition === undefined ||
+      condition_ast === undefined ||
+      variables === undefined
+    ) {
+      return 1;
     }
 
-    const ast = JSON.parse(condition_ast)
+    const ast = JSON.parse(condition_ast);
 
     const branchDistance = new BranchDistance();
 
-    let [distance, direct] = branchDistance.resolve(ast, variables)
+    const { value, direct } = branchDistance.resolve(ast, variables);
+    let distance = value;
 
     if (direct) {
-      distance = distance ? 0 : 1
+      distance = distance ? 0 : 1;
     }
 
     if (!target) {
-      distance = 1 - distance
+      distance = 1 - distance;
     }
 
     if (distance > 1 || distance < 0) {
-      throw new Error("Invalid distance!")
+      throw new Error("Invalid distance!");
     }
 
-    return distance
+    return distance;
   }
 
-  evaluate(ast: any, variables: any): [any, boolean] {
+  // eslint-disable-next-line
+  evaluate(
+    ast: t.Node,
+    variables: Record<string, unknown>
+  ): { value: any; direct: boolean } {
     if (!this[ast.type]) {
-      throw new Error(`Unimplemented ast type ${ast.type}`)
+      throw new Error(`Unimplemented ast type ${ast.type}`);
     }
 
-    return this[ast.type](ast, variables)
+    return this[ast.type](ast, variables);
   }
 
-  resolve(ast, variables: any): [any, boolean] {
+  // eslint-disable-next-line
+  resolve(
+    ast: t.Node,
+    variables: Record<string, unknown>
+  ): { value: any; direct: boolean } {
     switch (ast.type) {
       case "StringLiteral":
       case "NumericLiteral":
       case "BooleanLiteral":
-      case "TemplateLiteral":
+        return { value: ast.value, direct: true };
       case "RegExpLiteral":
-        return [ast.value, true]
+        return { value: ast.pattern, direct: true };
       case "NullLiteral":
-        return [null, true]
+        return { value: null, direct: true };
 
       case "Identifier":
-        const value = variables[ast.name]
         // TODO check if this is actually primitive?
-        return [value, true]
-      case "UpdateExpression":
-        return this.resolve(ast.argument, variables)
+        return { value: variables[ast.name], direct: true };
 
       case "MemberExpression":
-        if (ast.object.type === 'Identifier' && ast.property.type === 'Identifier') {
-          const value = variables[`${ast.object.name}.${ast.property.name}`]
-          return [value, true]
+        if (
+          ast.object.type === "Identifier" &&
+          ast.property.type === "Identifier"
+        ) {
+          const value = variables[`${ast.object.name}.${ast.property.name}`];
+          return { value: value, direct: true };
         }
 
       case "ThisExpression":
@@ -87,324 +101,245 @@ export class BranchDistance {
       // TODO the result would be cool but functions that alter state (side-effects) ruin the idea
       case "AssignmentExpression":
       case "CallExpression":
-        return [undefined, true]
+        return { value: undefined, direct: true };
     }
 
-    return this.evaluate(ast, variables)
+    return this.evaluate(ast, variables);
   }
 
-  UnaryExpression(ast: any, variables: any): [any, boolean] {
-    let [arg, direct] = this.resolve(ast.argument, variables)
+  // eslint-disable-next-line
+  UpdateExpression(
+    ast: t.UpdateExpression,
+    variables: Record<string, unknown>
+  ): { value: number; direct: boolean } {
+    const { value, direct } = this.resolve(ast.argument, variables);
+    if (direct) {
+      switch (ast.operator) {
+        // TODO postfix/prefix
+        case "++":
+          return { value: value + 1, direct: true };
+        case "--":
+          return { value: value - 1, direct: true };
+      }
+    }
+
+    throw new Error(
+      `Unknown update operator: "${ast.operator}" for direct=${direct}`
+    );
+  }
+
+  // eslint-disable-next-line
+  UnaryExpression(
+    ast: t.UnaryExpression,
+    variables: Record<string, unknown>
+  ): { value: any; direct: boolean } {
+    const { value, direct } = this.resolve(ast.argument, variables);
 
     if (direct) {
       switch (ast.operator) {
         case "!":
-          if (typeof arg === 'number') {
-            return [BranchDistance.normalize(Math.abs(0 - arg)), false]
+          if (typeof value === "number") {
+            return {
+              value: BranchDistance.normalize(Math.abs(0 - value)),
+              direct: false,
+            };
           } else {
-            return [arg ? 1 : 0, false]
+            return { value: value ? 1 : 0, direct: false };
           }
         case "typeof":
-          return [typeof arg, true]
+          return { value: typeof value, direct: true };
         case "-":
-          return [-arg, true]
+          return { value: -value, direct: true };
         case "+":
-          return [+arg, true]
-
-        // TODO postfix/prefix
-        case "++":
-          return [++arg, true]
-        case "--":
-          return [--arg, true]
-
+          return { value: +value, direct: true };
         case "~":
-          return [~arg, true]
+          return { value: ~value, direct: true };
       }
     } else {
       switch (ast.operator) {
         case "!":
-          return [1 - arg, false]
-
+          return { value: 1 - value, direct: false };
       }
     }
 
-    throw new Error(`Unknown unary operator: "${ast.operator}" for direct=${direct}`)
+    throw new Error(
+      `Unknown unary operator: "${ast.operator}" for direct=${direct}`
+    );
   }
 
-  BinaryExpression(ast: any, variables: any): [any, boolean] {
-    const [left, lDirect] = this.resolve(ast.left, variables)
-    const [right, rDirect] = this.resolve(ast.right, variables)
+  // eslint-disable-next-line
+  BinaryExpression(
+    ast: t.BinaryExpression,
+    variables: Record<string, unknown>
+  ): { value: any; direct: boolean } {
+    // eslint-disable-next-line
+    const { value: left, direct: lDirect } = this.resolve(ast.left, variables);
+    // eslint-disable-next-line
+    const { value: right, direct: rDirect } = this.resolve(
+      ast.right,
+      variables
+    );
 
     if (!lDirect) {
-      throw new Error("left should be direct")
+      throw new Error("left should be direct");
     }
     if (!rDirect) {
-      throw new Error("right not be direct")
+      throw new Error("right not be direct");
     }
 
     switch (ast.operator) {
       // should both be direct : returns direct
       case "+":
-        return [left + right, true]
+        return { value: left + right, direct: true };
       case "-":
-        return [left - right, true]
+        return { value: left - right, direct: true };
       case "*":
-        return [left * right, true]
+        return { value: left * right, direct: true };
       case "/":
-        return [left / right, true]
+        return { value: left / right, direct: true };
       case "%":
-        return [left % right, true]
+        return { value: left % right, direct: true };
       case "**":
-        return [left ** right, true]
+        return { value: left ** right, direct: true };
 
       case "&":
-        return [left & right, true]
+        return { value: left & right, direct: true };
       case "|":
-        return [left | right, true]
+        return { value: left | right, direct: true };
       case "^":
-        return [left ^ right, true]
+        return { value: left ^ right, direct: true };
     }
 
     switch (ast.operator) {
       // should both be direct : returns not direct
 
       case "==":
-        // TODO
+      // TODO
       case "===":
-        if (typeof left === 'number' && typeof right === 'number') {
-          return [BranchDistance.normalize(Math.abs(left - right)), false]
-        } else if (typeof left === 'string' && typeof right === 'string') {
-          return [BranchDistance.normalize(this.editDistDP(left, right)), false]
-        } else if (typeof left === 'boolean' && typeof right === 'boolean') {
-          return [left === right ? 0 : 1, false]
+        if (typeof left === "number" && typeof right === "number") {
+          return {
+            value: BranchDistance.normalize(Math.abs(left - right)),
+            direct: false,
+          };
+        } else if (typeof left === "string" && typeof right === "string") {
+          return {
+            value: BranchDistance.normalize(this.editDistDP(left, right)),
+            direct: false,
+          };
+        } else if (typeof left === "boolean" && typeof right === "boolean") {
+          return { value: left === right ? 0 : 1, direct: false };
         } else {
           // TODO type difference?!
         }
-        return [left === right ? 0 : 1, false]
+        return { value: left === right ? 0 : 1, direct: false };
 
       case "!=":
-        // TODO
+      // TODO
       case "!==":
-        return [left !== right ? 0 : 1, false]
+        return { value: left !== right ? 0 : 1, direct: false };
       case "<":
-        return [left < right ? 0 : BranchDistance.normalize(left - right + this.K), false]
+        return {
+          value:
+            left < right ? 0 : BranchDistance.normalize(left - right + this.K),
+          direct: false,
+        };
       case "<=":
-        return [left <= right ? 0 : BranchDistance.normalize(left - right), false]
+        return {
+          value: left <= right ? 0 : BranchDistance.normalize(left - right),
+          direct: false,
+        };
       case ">":
-        return [left > right ? 0 : BranchDistance.normalize(right - left + this.K), false]
+        return {
+          value:
+            left > right ? 0 : BranchDistance.normalize(right - left + this.K),
+          direct: false,
+        };
       case ">=":
-        return [left >= right ? 0 : BranchDistance.normalize(right - left), false]
+        return {
+          value: left >= right ? 0 : BranchDistance.normalize(right - left),
+          direct: false,
+        };
 
       case "instanceof":
-        return [left instanceof right ? 0 : 1, false]
+        return { value: left instanceof right ? 0 : 1, direct: false };
       case "in":
-        return [left in right ? 0 : 1, false]
+        return { value: left in right ? 0 : 1, direct: false };
     }
 
-    throw new Error(`Unknown binary operator: ${ast.operator}`)
+    throw new Error(`Unknown binary operator: ${ast.operator}`);
   }
 
-  LogicalExpression(ast: any, variables: any): [number, boolean] {
-    let [left, lDirect] = this.resolve(ast.left, variables)
-    let [right, rDirect] = this.resolve(ast.right, variables)
+  // eslint-disable-next-line
+  LogicalExpression(
+    ast: t.LogicalExpression,
+    variables: Record<string, unknown>
+  ): { value: any; direct: boolean } {
+    // eslint-disable-next-line
+    let { value: left, direct: lDirect } = this.resolve(ast.left, variables);
+    // eslint-disable-next-line
+    let { value: right, direct: rDirect } = this.resolve(ast.right, variables);
 
-    const operator = ast.operator
+    const operator = ast.operator;
 
     if (lDirect) {
-      left = left ? 0 : 1
+      left = left ? 0 : 1;
     }
 
     if (rDirect) {
-      right = right ? 0 : 1
+      right = right ? 0 : 1;
     }
 
     switch (operator) {
       // should both NOT be direct : returns not direct
       case "&&":
-        return [BranchDistance.normalize(left + right), false]
+        return { value: BranchDistance.normalize(left + right), direct: false };
       case "||":
-        return [BranchDistance.normalize(Math.min(left, right)), false]
+        return {
+          value: BranchDistance.normalize(Math.min(left, right)),
+          direct: false,
+        };
       // case ""
     }
 
-    throw new Error(`Unknown logical operator: ${ast.operator}`)
+    throw new Error(`Unknown logical operator: ${ast.operator}`);
   }
 
   private editDistDP(str1: string, str2: string) {
-    const m = str1.length
-    const n = str2.length
-    const table = []
+    const m = str1.length;
+    const n = str2.length;
+    const table = [];
 
     for (let i = 0; i <= m; i++) {
-      table.push([])
+      table.push([]);
       for (let j = 0; j <= n; j++) {
-        table[i].push(0)
+        table[i].push(0);
       }
     }
 
     for (let i = 0; i <= m; i++) {
       for (let j = 0; j <= n; j++) {
         if (i == 0) {
-          table[i][j] = j
+          table[i][j] = j;
         } else if (j == 0) {
-          table[i][j] = i
+          table[i][j] = i;
         } else if (str1.charAt(i - 1) === str2.charAt(j - 1)) {
-          table[i][j] = table[i - 1][j - 1]
+          table[i][j] = table[i - 1][j - 1];
         } else {
-          table[i][j] = 1 + Math.min(table[i][j - 1], Math.min(table[i - 1][j], table[i - 1][j - 1]))
+          table[i][j] =
+            1 +
+            Math.min(
+              table[i][j - 1],
+              Math.min(table[i - 1][j], table[i - 1][j - 1])
+            );
         }
       }
     }
 
-    return table[m][n]
-  }
-
-  /**
-   * Calculate the branch distance
-   *
-   * @param opcode the opcode (the comparison operator)
-   * @param left the left values of the comparison (multiple execution traces)
-   * @param right the right values of the comparison (multiple execution traces)
-   * @param target the side of the branch you want to cover
-   */
-  public static branchDistanceNumeric(
-    opcode: string,
-    left: number[],
-    right: number[],
-    target: boolean
-  ) {
-    let branchDistance: number;
-
-    // TODO the SGT and SLT opcodes are for signed numbers
-    // look here: https://docs.soliditylang.org/en/v0.5.5/assembly.html
-
-    // TODO other opcodes
-
-    // TODO move this to the solidity project and make an abstraction of this class
-
-    switch (opcode) {
-      case "EQ":
-        if (target) {
-          branchDistance = this.equalNumeric(left, right);
-        } else {
-          branchDistance = this.notEqualNumeric(left, right);
-        }
-        break;
-      case "NEQ":
-        if (target) {
-          branchDistance = this.notEqualNumeric(left, right);
-        } else {
-          branchDistance = this.equalNumeric(left, right);
-        }
-        break;
-      case "GT":
-        if (target) {
-          branchDistance = this.greater(left, right);
-        } else {
-          branchDistance = this.smallerEqual(left, right);
-        }
-        break;
-      case "LT":
-        if (target) {
-          branchDistance = this.smaller(left, right);
-        } else {
-          branchDistance = this.greaterEqual(left, right);
-        }
-        break;
-      case "SGT":
-        if (target) {
-          branchDistance = this.greater(left, right);
-        } else {
-          branchDistance = this.smallerEqual(left, right);
-        }
-        break;
-      case "SLT":
-        if (target) {
-          branchDistance = this.smaller(left, right);
-        } else {
-          branchDistance = this.greaterEqual(left, right);
-        }
-        break;
-    }
-
-    return this.normalize(branchDistance);
+    return table[m][n];
   }
 
   private static normalize(x: number): number {
     // return 1 - Math.pow(1.001, -x)
     return x / (x + 1);
-  }
-
-  private static equalNumeric(left: number[], right: number[]): number {
-    let minimum = Number.MAX_VALUE;
-    for (let index = 0; index < left.length; index++) {
-      minimum = Math.min(minimum, Math.abs(left[index] - right[index]));
-    }
-    return minimum;
-  }
-
-  private static notEqualNumeric(left: number[], right: number[]): number {
-    let minimum = Number.MAX_VALUE;
-
-    for (let index = 0; index < left.length; index++) {
-      if (left[index] != right[index]) {
-        minimum = 0.0;
-      } else {
-        minimum = Math.min(minimum, 1.0);
-      }
-    }
-    return minimum;
-  }
-
-  private static greater(left: number[], right: number[]): number {
-    let minimum = Number.MAX_VALUE;
-
-    for (let index = 0; index < left.length; index++) {
-      if (left[index] > right[index]) {
-        minimum = 0.0;
-      } else {
-        minimum = Math.min(minimum, right[index] - left[index] + 1);
-      }
-    }
-    return minimum;
-  }
-
-  private static smallerEqual(left: number[], right: number[]): number {
-    let minimum = Number.MAX_VALUE;
-
-    for (let index = 0; index < left.length; index++) {
-      if (left[index] <= right[index]) {
-        minimum = 0.0;
-      } else {
-        minimum = Math.min(minimum, left[index] - right[index]);
-      }
-    }
-    return minimum;
-  }
-
-  private static greaterEqual(left: number[], right: number[]): number {
-    let minimum = Number.MAX_VALUE;
-
-    for (let index = 0; index < left.length; index++) {
-      if (left[index] >= right[index]) {
-        minimum = 0.0;
-      } else {
-        minimum = Math.min(minimum, right[index] - left[index]);
-      }
-    }
-    return minimum;
-  }
-
-  private static smaller(left: number[], right: number[]): number {
-    let minimum = Number.MAX_VALUE;
-
-    for (let index = 0; index < left.length; index++) {
-      if (left[index] < right[index]) {
-        minimum = 0.0;
-      } else {
-        minimum = Math.min(minimum, left[index] - right[index] + 1);
-      }
-    }
-    return minimum;
   }
 }
