@@ -1,9 +1,10 @@
-import { BranchObjectiveFunction, Encoding } from "../../../syntest-framework";
+import { BranchObjectiveFunction, Encoding } from "@syntest/core";
 import { BranchDistance } from "./BranchDistance";
-import { Node, NodeType } from "@syntest/framework";
+import { Node, NodeType } from "@syntest/core";
 
-export class JavaScriptBranchObjectiveFunction<T extends Encoding> extends BranchObjectiveFunction<T> {
-
+export class JavaScriptBranchObjectiveFunction<
+  T extends Encoding
+> extends BranchObjectiveFunction<T> {
   calculateDistance(encoding: T): number {
     const executionResult = encoding.getExecutionResult();
 
@@ -24,7 +25,7 @@ export class JavaScriptBranchObjectiveFunction<T extends Encoding> extends Branc
 
       if (!branchTrace) {
         // TODO fix this should not happen
-        return Number.MAX_VALUE
+        return Number.MAX_VALUE;
       }
 
       if (branchTrace.hits > 0) {
@@ -40,9 +41,9 @@ export class JavaScriptBranchObjectiveFunction<T extends Encoding> extends Branc
         return BranchDistance.branchDistance(
           oppositeBranch.condition,
           oppositeBranch.condition_ast,
-          oppositeBranch.variables,
+          <Record<string, unknown>>oppositeBranch.variables,
           this._type
-        )
+        );
       }
     }
 
@@ -66,8 +67,7 @@ export class JavaScriptBranchObjectiveFunction<T extends Encoding> extends Branc
         .filter(
           (trace) =>
             n.lines.includes(trace.line) &&
-            (trace.type === "branch" ||
-              trace.type === "function") &&
+            (trace.type === "branch" || trace.type === "function") &&
             trace.hits > 0
         );
       for (const trace of traces) {
@@ -97,9 +97,9 @@ export class JavaScriptBranchObjectiveFunction<T extends Encoding> extends Branc
       branchDistance = BranchDistance.branchDistance(
         closestHitNode.condition,
         oppositeBranch.condition_ast,
-        oppositeBranch.variables,
+        <Record<string, unknown>>oppositeBranch.variables,
         this._type
-      )
+      );
     }
 
     // add the distances
