@@ -2,7 +2,7 @@ import { TransformOptions, transformSync } from "@babel/core";
 import { Visitor } from "./Visitor";
 import { defaultBabelOptions } from "../configs/DefaultBabelConfig";
 import * as path from "path";
-import { Properties } from "@syntest/framework";
+import { Properties } from "@syntest/core";
 const globby = require('globby');
 
 export interface OutputObject {
@@ -11,8 +11,8 @@ export interface OutputObject {
 }
 
 export class Instrumenter {
-  instrument(code: string, filename: string) {
-    const options = JSON.parse(JSON.stringify(defaultBabelOptions)) ;
+  async instrument(code: string, filename: string) {
+    const options = JSON.parse(JSON.stringify(defaultBabelOptions));
 
     let output: OutputObject = {};
 
@@ -59,7 +59,7 @@ export class Instrumenter {
       return;
     }
 
-    const codeMap = transformSync(code, options);
+    const codeMap = await transformSync(code, options);
 
     if (!output || !output.fileCoverage) {
       return code;
