@@ -16,11 +16,7 @@
  * limitations under the License.
  */
 
-import {
-  getUserInterface,
-  Encoding,
-  Decoder, Properties,
-} from "@syntest/framework";
+import { getUserInterface, Encoding, Decoder, Properties } from "@syntest/core";
 import { JavaScriptTestCaseSampler } from "./sampling/JavaScriptTestCaseSampler";
 import { RootStatement } from "./statements/root/RootStatement";
 
@@ -45,11 +41,9 @@ export class JavaScriptTestCase extends Encoding {
   mutate(sampler: JavaScriptTestCaseSampler) {
     getUserInterface().debug(`Mutating test case: ${this._id}`);
     if (Properties.resample_gene_probability) {
-      return sampler.sample()
+      return sampler.sample();
     } else {
-      return new JavaScriptTestCase(
-        this._root.mutate(sampler, 0)
-      );
+      return new JavaScriptTestCase(this._root.mutate(sampler, 0));
     }
   }
 
