@@ -17,7 +17,7 @@
  */
 
 import * as chalk from "chalk";
-import {textSync} from "figlet";
+import { textSync } from "figlet";
 
 /**
  * This class contains standardized functions that create colored and formatted messages to be used by the user-interface.
@@ -70,10 +70,23 @@ export default class Messages {
     return "\n" + chalk.green(chalk.bold(`${header}`)) + "\n";
   }
 
+  // eslint-disable-next-line
   reportCoverage(name: string, report: any, bold = false): string {
     if (bold) {
-      return chalk.bold(`${String(report['branch']).padStart(20, ' ')} | ${String(report['statement']).padStart(20, ' ')} | ${String(report['function']).padStart(20, ' ')} | ${name}`)
+      return chalk.bold(
+        `${String(report["branch"]).padStart(20, " ")} | ${String(
+          report["statement"]
+        ).padStart(20, " ")} | ${String(report["function"]).padStart(
+          20,
+          " "
+        )} | ${name}`
+      );
     }
-    return `${String(report['branch']).padStart(20, ' ')} | ${String(report['statement']).padStart(20, ' ')} | ${String(report['function']).padStart(20, ' ')} | ${name}`
+    return `${String(report["branch"]).padStart(20, " ")} | ${String(
+      report["statement"]
+    ).padStart(20, " ")} | ${String(report["function"]).padStart(
+      20,
+      " "
+    )} | ${name}`;
   }
 }
