@@ -29,15 +29,16 @@ import {
   SearchTimeBudget,
   StatisticsCollector,
   TotalTimeBudget,
-} from "@syntest/framework";
+  Encoding,
+} from "@syntest/core";
 import { JavaScriptSubject } from "../search/JavaScriptSubject";
 
-export function collectInitialVariables(
-  collector: StatisticsCollector<any>,
+export function collectInitialVariables<T extends Encoding>(
+  collector: StatisticsCollector<T>,
   currentSubject: JavaScriptSubject,
   targetPath: string
 ) {
-  collector.recordVariable(RuntimeVariable.VERSION, 1);
+  collector.recordVariable(RuntimeVariable.VERSION, "1");
   collector.recordVariable(
     RuntimeVariable.CONFIGURATION,
     Properties.configuration
@@ -46,48 +47,48 @@ export function collectInitialVariables(
   collector.recordVariable(RuntimeVariable.SUBJECT, path.basename(targetPath));
   collector.recordVariable(
     RuntimeVariable.CONSTANT_POOL_ENABLED,
-    Properties.constant_pool
+    `${Properties.constant_pool}`
   );
   collector.recordVariable(RuntimeVariable.ALGORITHM, Properties.algorithm);
   collector.recordVariable(
     RuntimeVariable.TOTAL_OBJECTIVES,
-    currentSubject.getObjectives().length
+    `${currentSubject.getObjectives().length}`
   );
 }
 
-export function collectStatistics(
-  collector: StatisticsCollector<any>,
+export function collectStatistics<T extends Encoding>(
+  collector: StatisticsCollector<T>,
   currentSubject: JavaScriptSubject,
-  archive: Archive<any>,
-  totalTimeBudget: TotalTimeBudget<any>,
-  searchBudget: SearchTimeBudget<any>,
-  iterationBudget: IterationBudget<any>,
-  evaluationBudget: EvaluationBudget<any>
+  archive: Archive<T>,
+  totalTimeBudget: TotalTimeBudget<T>,
+  searchBudget: SearchTimeBudget<T>,
+  iterationBudget: IterationBudget<T>,
+  evaluationBudget: EvaluationBudget<T>
 ) {
   collector.recordVariable(
     RuntimeVariable.COVERED_OBJECTIVES,
-    archive.getObjectives().length
+    `${archive.getObjectives().length}`
   );
 
   collector.recordVariable(
     RuntimeVariable.INITIALIZATION_TIME,
-    totalTimeBudget.getUsedBudget() - searchBudget.getUsedBudget()
+    `${totalTimeBudget.getUsedBudget() - searchBudget.getUsedBudget()}`
   );
   collector.recordVariable(
     RuntimeVariable.SEARCH_TIME,
-    searchBudget.getUsedBudget()
+    `${searchBudget.getUsedBudget()}`
   );
   collector.recordVariable(
     RuntimeVariable.TOTAL_TIME,
-    totalTimeBudget.getUsedBudget()
+    `${totalTimeBudget.getUsedBudget()}`
   );
   collector.recordVariable(
     RuntimeVariable.ITERATIONS,
-    iterationBudget.getUsedBudget()
+    `${iterationBudget.getUsedBudget()}`
   );
   collector.recordVariable(
     RuntimeVariable.EVALUATIONS,
-    evaluationBudget.getUsedBudget()
+    `${evaluationBudget.getUsedBudget()}`
   );
 
   const numOfExceptions = archive
@@ -96,18 +97,23 @@ export function collectStatistics(
       (objective) => objective instanceof ExceptionObjectiveFunction
     ).length;
 
-  collector.recordVariable(RuntimeVariable.COVERED_EXCEPTIONS, numOfExceptions);
+  collector.recordVariable(
+    RuntimeVariable.COVERED_EXCEPTIONS,
+    `${numOfExceptions}`
+  );
 
   collector.recordVariable(
     RuntimeVariable.COVERAGE,
-    (archive.getObjectives().length - numOfExceptions) /
+    `${
+      (archive.getObjectives().length - numOfExceptions) /
       currentSubject.getObjectives().length
+    }`
   );
 }
 
-export function collectCoverageData(
-  collector: StatisticsCollector<any>,
-  archive: Archive<any>,
+export function collectCoverageData<T extends Encoding>(
+  collector: StatisticsCollector<T>,
+  archive: Archive<T>,
   objectiveType: string
 ): void {
   const total = new Set();
@@ -118,21 +124,17 @@ export function collectCoverageData(
     const result: ExecutionResult = test.getExecutionResult();
     // TODO this does not work when there are files with the same name in different directories!!
     const paths = key.getSubject().path.split("/");
-    const fileName = paths[paths.length - 1]
+    const fileName = paths[paths.length - 1];
 
     result
       .getTraces()
       .filter((element) => element.type.includes(objectiveType))
       .filter((element) => element.path.includes(fileName))
       .forEach((current) => {
-        total.add(
-          current.id + "_" + current.branchType
-        );
+        total.add(current.id + "_" + current.branchType);
 
         if (current.hits > 0)
-          covered.add(
-            current.id + "_" + current.branchType
-          );
+          covered.add(current.id + "_" + current.branchType);
       });
   }
 
@@ -141,32 +143,38 @@ export function collectCoverageData(
       {
         collector.recordVariable(
           RuntimeVariable.COVERED_BRANCHES,
-          covered.size
+          `${covered.size}`
         );
-        collector.recordVariable(RuntimeVariable.TOTAL_BRANCHES, total.size);
+        collector.recordVariable(
+          RuntimeVariable.TOTAL_BRANCHES,
+          `${total.size}`
+        );
 
         if (total.size > 0.0) {
           collector.recordVariable(
             RuntimeVariable.BRANCH_COVERAGE,
-            covered.size / total.size
+            `${covered.size / total.size}`
           );
         } else {
-          collector.recordVariable(RuntimeVariable.BRANCH_COVERAGE, 0);
+          collector.recordVariable(RuntimeVariable.BRANCH_COVERAGE, "0");
         }
       }
       break;
     case "statement":
       {
-        collector.recordVariable(RuntimeVariable.COVERED_LINES, covered.size);
-        collector.recordVariable(RuntimeVariable.TOTAL_LINES, total.size);
+        collector.recordVariable(
+          RuntimeVariable.COVERED_LINES,
+          `${covered.size}`
+        );
+        collector.recordVariable(RuntimeVariable.TOTAL_LINES, `${total.size}`);
 
         if (total.size > 0.0) {
           collector.recordVariable(
             RuntimeVariable.LINE_COVERAGE,
-            covered.size / total.size
+            `${covered.size / total.size}`
           );
         } else {
-          collector.recordVariable(RuntimeVariable.LINE_COVERAGE, 0);
+          collector.recordVariable(RuntimeVariable.LINE_COVERAGE, "0");
         }
       }
       break;
@@ -174,17 +182,20 @@ export function collectCoverageData(
       {
         collector.recordVariable(
           RuntimeVariable.COVERED_FUNCTIONS,
-          covered.size
+          `${covered.size}`
         );
-        collector.recordVariable(RuntimeVariable.TOTAL_FUNCTIONS, total.size);
+        collector.recordVariable(
+          RuntimeVariable.TOTAL_FUNCTIONS,
+          `${total.size}`
+        );
 
         if (total.size > 0.0) {
           collector.recordVariable(
             RuntimeVariable.FUNCTION_COVERAGE,
-            covered.size / total.size
+            `${covered.size / total.size}`
           );
         } else {
-          collector.recordVariable(RuntimeVariable.FUNCTION_COVERAGE, 0);
+          collector.recordVariable(RuntimeVariable.FUNCTION_COVERAGE, "0");
         }
       }
       break;
